@@ -24,6 +24,7 @@ import static android.provider.Settings.ACTION_APPLICATION_DETAILS_SETTINGS;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
+import android.content.ActivityNotFoundException;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.ApplicationInfo;
@@ -33,8 +34,6 @@ import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Build;
-import android.os.Handler;
-import android.os.HandlerThread;
 import android.text.Spannable;
 import android.text.SpannableStringBuilder;
 import android.text.TextUtils;
@@ -49,6 +48,7 @@ import android.widget.CompoundButton;
 import android.widget.Filter;
 import android.widget.Filterable;
 import android.widget.ImageView;
+import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -61,80 +61,91 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.bumptech.glide.request.target.CustomTarget;
 import com.bumptech.glide.request.transition.Transition;
 import com.google.android.material.checkbox.MaterialCheckBox;
-import com.google.android.material.dialog.MaterialAlertDialogBuilder;
-import com.google.android.material.snackbar.Snackbar;
 
 import org.lsposed.lspd.models.Application;
 import org.lsposed.manager.App;
 import org.lsposed.manager.BuildConfig;
 import org.lsposed.manager.ConfigManager;
 import org.lsposed.manager.R;
+import org.lsposed.manager.databinding.ItemMasterSwitchBinding;
 import org.lsposed.manager.databinding.ItemModuleBinding;
+import org.lsposed.manager.ui.dialog.BlurBehindDialogBuilder;
 import org.lsposed.manager.ui.fragment.AppListFragment;
 import org.lsposed.manager.ui.fragment.CompileDialogFragment;
+import org.lsposed.manager.ui.widget.EmptyStateRecyclerView;
 import org.lsposed.manager.util.GlideApp;
 import org.lsposed.manager.util.ModuleUtil;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Objects;
-import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import rikka.core.util.ResourceUtils;
-import rikka.widget.switchbar.SwitchBar;
+import rikka.material.app.LocaleDelegate;
+import rikka.widget.mainswitchbar.MainSwitchBar;
+import rikka.widget.mainswitchbar.OnMainSwitchChangeListener;
 
-@SuppressLint("NotifyDataSetChanged")
-public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> implements Filterable {
+public class ScopeAdapter extends EmptyStateRecyclerView.EmptyStateAdapter<ScopeAdapter.ViewHolder> implements Filterable {
 
     private final Activity activity;
     private final AppListFragment fragment;
     private final PackageManager pm;
     private final SharedPreferences preferences;
-    private final HandlerThread handlerThread = new HandlerThread("appList");
-    private final Handler loadAppListHandler;
     private final ModuleUtil moduleUtil;
 
     private final ModuleUtil.InstalledModule module;
 
-    private final HashSet<ApplicationWithEquals> recommendedList = new HashSet<>();
-    private final HashSet<ApplicationWithEquals> checkedList = new HashSet<>();
-    private final ConcurrentLinkedQueue<AppInfo> searchList = new ConcurrentLinkedQueue<>();
-    private final List<AppInfo> showList = new ArrayList<>();
+    private Set<ApplicationWithEquals> recommendedList = new HashSet<>();
+    private Set<ApplicationWithEquals> checkedList = new HashSet<>();
+    private List<AppInfo> searchList = new ArrayList<>();
+    private List<AppInfo> showList = new ArrayList<>();
+    private List<String> denyList = new ArrayList<>();
 
-    private final SwitchBar.OnCheckedChangeListener switchBarOnCheckedChangeListener = new SwitchBar.OnCheckedChangeListener() {
+    public RecyclerView.Adapter<RecyclerView.ViewHolder> switchAdaptor = new RecyclerView.Adapter<>() {
+        @NonNull
         @Override
-        public boolean onCheckedChanged(SwitchBar view, boolean isChecked) {
-            if (!moduleUtil.setModuleEnabled(module.packageName, isChecked)) {
-                return false;
-            }
+        public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+            return new RecyclerView.ViewHolder(ItemMasterSwitchBinding.inflate(activity.getLayoutInflater(), parent, false).masterSwitch) {
+            };
+        }
+
+        @Override
+        public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position) {
+            var mainSwitchBar = (MainSwitchBar) holder.itemView;
+            mainSwitchBar.setChecked(enabled);
+            mainSwitchBar.addOnSwitchChangeListener(switchBarOnCheckedChangeListener);
+        }
+
+        @Override
+        public int getItemCount() {
+            return 1;
+        }
+    };
+
+    private final OnMainSwitchChangeListener switchBarOnCheckedChangeListener = new OnMainSwitchChangeListener() {
+        @Override
+        public void onSwitchChanged(Switch view, boolean isChecked) {
             enabled = isChecked;
-            notifyDataSetChanged();
-            return true;
-        }
-    };
-    private final Runnable dataReadyRunnable = new Runnable() {
-        @Override
-        public void run() {
-            synchronized (this) {
-                if (fragment == null || fragment.binding == null) {
-                    return;
-                }
-                fragment.binding.progress.setIndeterminate(false);
-                fragment.binding.swipeRefreshLayout.setRefreshing(false);
-                String queryStr = fragment.searchView != null ? fragment.searchView.getQuery().toString() : "";
-                getFilter().filter(queryStr);
-                this.notify();
+            if (!moduleUtil.setModuleEnabled(module.packageName, isChecked)) {
+                view.setChecked(!isChecked);
+                enabled = !isChecked;
             }
+            var tmpChkList = new HashSet<>(checkedList);
+            if (isChecked && !tmpChkList.isEmpty() && !ConfigManager.setModuleScope(module.packageName, module.legacy, tmpChkList)) {
+                view.setChecked(false);
+                enabled = false;
+            }
+            fragment.runOnUiThread(ScopeAdapter.this::notifyDataSetChanged);
         }
     };
 
-    private ApplicationInfo selectedInfo;
-    private boolean refreshing = false;
+    private ApplicationInfo selectedApplicationInfo;
+    private boolean isLoaded = false;
     private boolean enabled = true;
 
     public ScopeAdapter(AppListFragment fragment, ModuleUtil.InstalledModule module) {
@@ -142,8 +153,6 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         this.activity = fragment.requireActivity();
         this.module = module;
         moduleUtil = ModuleUtil.getInstance();
-        handlerThread.start();
-        loadAppListHandler = new Handler(handlerThread.getLooper());
         preferences = App.getPreferences();
         pm = activity.getPackageManager();
     }
@@ -154,15 +163,20 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         return new ViewHolder(ItemModuleBinding.inflate(activity.getLayoutInflater(), parent, false));
     }
 
-    private boolean shouldHideApp(PackageInfo info, ApplicationWithEquals app) {
-        if (info.packageName.equals("android")) {
+    private boolean shouldHideApp(PackageInfo info, ApplicationWithEquals app, HashSet<ApplicationWithEquals> tmpChkList) {
+        if (info.packageName.equals("system")) {
             return false;
         }
-        if (checkedList.contains(app)) {
+        if (tmpChkList.contains(app)) {
             return false;
+        }
+        if (preferences.getBoolean("filter_denylist", false)) {
+            if (denyList.contains(info.packageName)) {
+                return true;
+            }
         }
         if (preferences.getBoolean("filter_modules", true)) {
-            if (info.applicationInfo.metaData != null && info.applicationInfo.metaData.containsKey("xposedminversion")) {
+            if (ModuleUtil.getInstance().getModule(info.packageName, info.applicationInfo.uid / App.PER_USER_RANGE) != null) {
                 return true;
             }
         }
@@ -175,18 +189,15 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
                 return true;
             }
         }
-        if ((info.applicationInfo.flags & ApplicationInfo.FLAG_HAS_CODE) == 0) {
-            return true;
-        }
         return preferences.getBoolean("filter_system_apps", true) && (info.applicationInfo.flags & ApplicationInfo.FLAG_SYSTEM) != 0;
     }
 
-    private void sortApps(List<AppInfo> list) {
+    private int sortApps(AppInfo x, AppInfo y) {
         Comparator<PackageInfo> comparator = AppHelper.getAppListComparator(preferences.getInt("list_sort", 0), pm);
         Comparator<AppInfo> frameworkComparator = (a, b) -> {
-            if (a.packageName.equals("android") == b.packageName.equals("android")) {
+            if (a.packageName.equals("system") == b.packageName.equals("system")) {
                 return comparator.compare(a.packageInfo, b.packageInfo);
-            } else if (a.packageName.equals("android")) {
+            } else if (a.packageName.equals("system")) {
                 return -1;
             } else {
                 return 1;
@@ -203,40 +214,52 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
                 return 1;
             }
         };
-        list.sort((a, b) -> {
-            boolean aChecked = checkedList.contains(a.application);
-            boolean bChecked = checkedList.contains(b.application);
-            if (aChecked == bChecked) {
-                return recommendedComparator.compare(a, b);
-            } else if (aChecked) {
-                return -1;
-            } else {
-                return 1;
-            }
-        });
+        boolean aChecked = checkedList.contains(x.application);
+        boolean bChecked = checkedList.contains(y.application);
+        if (aChecked == bChecked) {
+            return recommendedComparator.compare(x, y);
+        } else if (aChecked) {
+            return -1;
+        } else {
+            return 1;
+        }
     }
 
     private void checkRecommended() {
-        checkedList.removeIf(i -> i.userId == module.userId);
-        checkedList.addAll(recommendedList);
-        ConfigManager.setModuleScope(module.packageName, checkedList);
+        if (!enabled) {
+            fragment.showHint(R.string.module_is_not_activated_yet, false);
+            return;
+        }
+        fragment.runAsync(() -> {
+            var tmpChkList = new HashSet<>(checkedList);
+            tmpChkList.removeIf(i -> i.userId == module.userId);
+            tmpChkList.addAll(recommendedList);
+            ConfigManager.setModuleScope(module.packageName, module.legacy, tmpChkList);
+            checkedList = tmpChkList;
+            fragment.runOnUiThread(this::notifyDataSetChanged);
+        });
+    }
+
+    @SuppressLint("NotifyDataSetChanged")
+    private void setLoaded(List<AppInfo> list, boolean loaded) {
+        fragment.runOnUiThread(() -> {
+            if (list != null) showList = list;
+            isLoaded = loaded;
+            notifyDataSetChanged();
+        });
     }
 
     public boolean onOptionsItemSelected(MenuItem item) {
         int itemId = item.getItemId();
         if (itemId == R.id.use_recommended) {
             if (!checkedList.isEmpty()) {
-                new MaterialAlertDialogBuilder(activity)
+                new BlurBehindDialogBuilder(activity, R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons)
                         .setMessage(R.string.use_recommended_message)
-                        .setPositiveButton(android.R.string.ok, (dialog, which) -> {
-                            checkRecommended();
-                            notifyDataSetChanged();
-                        })
+                        .setPositiveButton(android.R.string.ok, (dialog, which) -> checkRecommended())
                         .setNegativeButton(android.R.string.cancel, null)
                         .show();
             } else {
                 checkRecommended();
-                notifyDataSetChanged();
             }
             return true;
         } else if (itemId == R.id.item_filter_system) {
@@ -248,22 +271,27 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         } else if (itemId == R.id.item_filter_modules) {
             item.setChecked(!item.isChecked());
             preferences.edit().putBoolean("filter_modules", item.isChecked()).apply();
-        } else if (itemId == R.id.menu_launch) {
-            Intent launchIntent = AppHelper.getSettingsIntent(module.packageName, module.userId);
-            if (launchIntent != null) {
-                ConfigManager.startActivityAsUserWithFeature(launchIntent, module.userId);
-            } else {
-                Snackbar.make(fragment.binding.snackbar, R.string.module_no_ui, Snackbar.LENGTH_LONG).show();
-            }
-            return true;
+        } else if (itemId == R.id.item_filter_denylist) {
+            item.setChecked(!item.isChecked());
+            preferences.edit().putBoolean("filter_denylist", item.isChecked()).apply();
         } else if (itemId == R.id.backup) {
             LocalDateTime now = LocalDateTime.now();
-            fragment.backupLauncher.launch(String.format(Locale.ROOT,
-                    "%s_%s.lsp", module.getAppName(), now.toString()));
-            return true;
+            try {
+                fragment.backupLauncher.launch(String.format(LocaleDelegate.getDefaultLocale(),
+                        "%s_%s.lsp", module.getAppName(), now.toString()));
+                return true;
+            } catch (ActivityNotFoundException e) {
+                fragment.showHint(R.string.enable_documentui, true);
+                return false;
+            }
         } else if (itemId == R.id.restore) {
-            fragment.restoreLauncher.launch(new String[]{"*/*"});
-            return true;
+            try {
+                fragment.restoreLauncher.launch(new String[]{"*/*"});
+                return true;
+            } catch (ActivityNotFoundException e) {
+                fragment.showHint(R.string.enable_documentui, true);
+                return false;
+            }
         } else if (!AppHelper.onOptionsItemSelected(item, preferences)) {
             return false;
         }
@@ -272,18 +300,18 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
     }
 
     public boolean onContextItemSelected(@NonNull MenuItem item) {
-        ApplicationInfo info = selectedInfo;
+        var info = selectedApplicationInfo;
         if (info == null) {
             return false;
         }
         int itemId = item.getItemId();
         if (itemId == R.id.menu_launch) {
-            Intent launchIntent = AppHelper.getLaunchIntentForPackage(info.packageName, info.uid / 100000);
+            Intent launchIntent = AppHelper.getLaunchIntentForPackage(info.packageName, info.uid / App.PER_USER_RANGE);
             if (launchIntent != null) {
                 ConfigManager.startActivityAsUserWithFeature(launchIntent, module.userId);
             }
         } else if (itemId == R.id.menu_compile_speed) {
-            CompileDialogFragment.speed(fragment.getChildFragmentManager(), info, fragment.binding.snackbar);
+            CompileDialogFragment.speed(fragment.getChildFragmentManager(), info);
         } else if (itemId == R.id.menu_other_app) {
             var intent = new Intent(Intent.ACTION_SHOW_APP_INFO);
             intent.putExtra(Intent.EXTRA_PACKAGE_NAME, module.packageName);
@@ -292,10 +320,14 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         } else if (itemId == R.id.menu_app_info) {
             ConfigManager.startActivityAsUserWithFeature(new Intent(ACTION_APPLICATION_DETAILS_SETTINGS, Uri.fromParts("package", info.packageName, null)), module.userId);
         } else if (itemId == R.id.menu_force_stop) {
-            if (info.packageName.equals("android")) {
-                ConfigManager.reboot(false);
+            if (info.packageName.equals("system")) {
+                new BlurBehindDialogBuilder(activity, R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons)
+                        .setTitle(R.string.reboot)
+                        .setPositiveButton(android.R.string.ok, (dialog, which) -> ConfigManager.reboot())
+                        .setNegativeButton(android.R.string.cancel, null)
+                        .show();
             } else {
-                new MaterialAlertDialogBuilder(activity)
+                new BlurBehindDialogBuilder(activity, R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons)
                         .setTitle(R.string.force_stop_dlg_title)
                         .setMessage(R.string.force_stop_dlg_text)
                         .setPositiveButton(android.R.string.ok, (dialog, which) -> ConfigManager.forceStopPackage(info.packageName, info.uid / 100000))
@@ -309,10 +341,6 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
     }
 
     public void onPrepareOptionsMenu(@NonNull Menu menu) {
-        Intent intent = AppHelper.getSettingsIntent(module.packageName, module.userId);
-        if (intent == null) {
-            menu.removeItem(R.id.menu_launch);
-        }
         List<String> scopeList = module.getScopeList();
         if (scopeList == null || scopeList.isEmpty()) {
             menu.removeItem(R.id.use_recommended);
@@ -320,72 +348,80 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         menu.findItem(R.id.item_filter_system).setChecked(preferences.getBoolean("filter_system_apps", true));
         menu.findItem(R.id.item_filter_games).setChecked(preferences.getBoolean("filter_games", true));
         menu.findItem(R.id.item_filter_modules).setChecked(preferences.getBoolean("filter_modules", true));
+        menu.findItem(R.id.item_filter_denylist).setChecked(preferences.getBoolean("filter_denylist", false));
         switch (preferences.getInt("list_sort", 0)) {
-            case 7:
+            case 7 -> {
                 menu.findItem(R.id.item_sort_by_update_time).setChecked(true);
                 menu.findItem(R.id.reverse).setChecked(true);
-                break;
-            case 6:
-                menu.findItem(R.id.item_sort_by_update_time).setChecked(true);
-                break;
-            case 5:
+            }
+            case 6 -> menu.findItem(R.id.item_sort_by_update_time).setChecked(true);
+            case 5 -> {
                 menu.findItem(R.id.item_sort_by_install_time).setChecked(true);
                 menu.findItem(R.id.reverse).setChecked(true);
-                break;
-            case 4:
-                menu.findItem(R.id.item_sort_by_install_time).setChecked(true);
-                break;
-            case 3:
+            }
+            case 4 -> menu.findItem(R.id.item_sort_by_install_time).setChecked(true);
+            case 3 -> {
                 menu.findItem(R.id.item_sort_by_package_name).setChecked(true);
                 menu.findItem(R.id.reverse).setChecked(true);
-                break;
-            case 2:
-                menu.findItem(R.id.item_sort_by_package_name).setChecked(true);
-                break;
-            case 1:
+            }
+            case 2 -> menu.findItem(R.id.item_sort_by_package_name).setChecked(true);
+            case 1 -> {
                 menu.findItem(R.id.item_sort_by_name).setChecked(true);
                 menu.findItem(R.id.reverse).setChecked(true);
-                break;
-            case 0:
-                menu.findItem(R.id.item_sort_by_name).setChecked(true);
-                break;
+            }
+            case 0 -> menu.findItem(R.id.item_sort_by_name).setChecked(true);
         }
     }
 
     @Override
+    public void onViewRecycled(@NonNull ViewHolder holder) {
+        if (holder.checkbox != null) {
+            holder.checkbox.setOnCheckedChangeListener(null);
+        }
+        super.onViewRecycled(holder);
+    }
+
+    @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        holder.root.setAlpha(enabled ? 1.0f : .5f);
         AppInfo appInfo = showList.get(position);
-        boolean android = appInfo.packageName.equals("android");
+        boolean deny = denyList.contains(appInfo.packageName);
+        holder.root.setAlpha(!deny && enabled ? 1.0f : .5f);
+        boolean system = appInfo.packageName.equals("system");
         CharSequence appName;
-        int userId = appInfo.applicationInfo.uid / 100000;
-        appName = android ? activity.getString(R.string.android_framework) : appInfo.label;
+        int userId = appInfo.applicationInfo.uid / App.PER_USER_RANGE;
+        appName = system ? activity.getString(R.string.android_framework) : appInfo.label;
         holder.appName.setText(appName);
-        GlideApp.with(holder.appIcon)
-                .load(appInfo.packageInfo)
-                .into(new CustomTarget<Drawable>() {
-                    @Override
-                    public void onResourceReady(@NonNull Drawable resource, @Nullable Transition<? super Drawable> transition) {
-                        holder.appIcon.setImageDrawable(resource);
-                    }
+        GlideApp.with(holder.appIcon).load(appInfo.packageInfo).into(new CustomTarget<Drawable>() {
+            @Override
+            public void onResourceReady(@NonNull Drawable resource, @Nullable Transition<? super Drawable> transition) {
+                holder.appIcon.setImageDrawable(resource);
+            }
 
-                    @Override
-                    public void onLoadCleared(@Nullable Drawable placeholder) {
+            @Override
+            public void onLoadCleared(@Nullable Drawable placeholder) {
 
-                    }
+            }
 
-                    @Override
-                    public void onLoadFailed(@Nullable Drawable errorDrawable) {
-                        holder.appIcon.setImageDrawable(pm.getDefaultActivityIcon());
-                    }
-                });
-        SpannableStringBuilder sb = new SpannableStringBuilder(android ? "" : activity.getString(R.string.app_description, appInfo.packageName, appInfo.packageInfo.versionName));
-        holder.appDescription.setVisibility(View.VISIBLE);
+            @Override
+            public void onLoadFailed(@Nullable Drawable errorDrawable) {
+                holder.appIcon.setImageDrawable(pm.getDefaultActivityIcon());
+            }
+        });
+        if (system) {
+            //noinspection SetTextI18n
+            holder.appPackageName.setText("system");
+            holder.appVersionName.setVisibility(View.GONE);
+        } else {
+            holder.appVersionName.setVisibility(View.VISIBLE);
+            holder.appPackageName.setText(appInfo.packageName);
+        }
+        holder.appPackageName.setVisibility(View.VISIBLE);
+        holder.appVersionName.setText(activity.getString(R.string.app_version, appInfo.packageInfo.versionName));
+        var sb = new SpannableStringBuilder();
         if (!recommendedList.isEmpty() && recommendedList.contains(appInfo.application)) {
-            if (!android) sb.append("\n");
             String recommended = activity.getString(R.string.requested_by_module);
             sb.append(recommended);
-            final ForegroundColorSpan foregroundColorSpan = new ForegroundColorSpan(ResourceUtils.resolveColor(activity.getTheme(), androidx.appcompat.R.attr.colorAccent));
+            final ForegroundColorSpan foregroundColorSpan = new ForegroundColorSpan(ResourceUtils.resolveColor(activity.getTheme(), com.google.android.material.R.attr.colorPrimary));
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
                 final TypefaceSpan typefaceSpan = new TypefaceSpan(Typeface.create("sans-serif-medium", Typeface.NORMAL));
                 sb.setSpan(typefaceSpan, sb.length() - recommended.length(), sb.length(), Spannable.SPAN_INCLUSIVE_INCLUSIVE);
@@ -394,10 +430,27 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
                 sb.setSpan(styleSpan, sb.length() - recommended.length(), sb.length(), Spannable.SPAN_INCLUSIVE_INCLUSIVE);
             }
             sb.setSpan(foregroundColorSpan, sb.length() - recommended.length(), sb.length(), Spannable.SPAN_INCLUSIVE_INCLUSIVE);
-        } else if (android) {
-            holder.appDescription.setVisibility(View.GONE);
         }
-        holder.appDescription.setText(sb);
+        if (deny) {
+            if (sb.length() != 0) sb.append("\n");
+            String denylist = activity.getString(R.string.deny_list_info);
+            sb.append(denylist);
+            final ForegroundColorSpan foregroundColorSpan = new ForegroundColorSpan(ResourceUtils.resolveColor(activity.getTheme(), com.google.android.material.R.attr.colorError));
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                final TypefaceSpan typefaceSpan = new TypefaceSpan(Typeface.create("sans-serif-medium", Typeface.NORMAL));
+                sb.setSpan(typefaceSpan, sb.length() - denylist.length(), sb.length(), Spannable.SPAN_INCLUSIVE_INCLUSIVE);
+            } else {
+                final StyleSpan styleSpan = new StyleSpan(Typeface.BOLD);
+                sb.setSpan(styleSpan, sb.length() - denylist.length(), sb.length(), Spannable.SPAN_INCLUSIVE_INCLUSIVE);
+            }
+            sb.setSpan(foregroundColorSpan, sb.length() - denylist.length(), sb.length(), Spannable.SPAN_INCLUSIVE_INCLUSIVE);
+        }
+        if (sb.length() == 0) {
+            holder.hint.setVisibility(View.GONE);
+        } else {
+            holder.hint.setText(sb);
+            holder.hint.setVisibility(View.VISIBLE);
+        }
 
         holder.itemView.setOnCreateContextMenuListener((menu, v, menuInfo) -> {
             activity.getMenuInflater().inflate(R.menu.menu_app_item, menu);
@@ -406,22 +459,24 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
             if (launchIntent == null) {
                 menu.removeItem(R.id.menu_launch);
             }
-            if (android) {
+            if (system) {
                 menu.findItem(R.id.menu_force_stop).setTitle(R.string.reboot);
                 menu.removeItem(R.id.menu_compile_speed);
+                menu.removeItem(R.id.menu_other_app);
+                menu.removeItem(R.id.menu_app_info);
             }
         });
 
-        holder.checkbox.setOnCheckedChangeListener(null);
         holder.checkbox.setChecked(checkedList.contains(appInfo.application));
 
         holder.checkbox.setOnCheckedChangeListener((v, isChecked) -> onCheckedChange(v, isChecked, appInfo));
+
         holder.itemView.setOnClickListener(v -> {
             if (enabled) holder.checkbox.toggle();
         });
         holder.itemView.setOnLongClickListener(v -> {
             fragment.searchView.clearFocus();
-            selectedInfo = appInfo.applicationInfo;
+            selectedApplicationInfo = appInfo.applicationInfo;
             return false;
         });
     }
@@ -429,7 +484,7 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
     @Override
     public long getItemId(int position) {
         PackageInfo info = showList.get(position).packageInfo;
-        return (info.packageName + "!" + info.applicationInfo.uid / 100000).hashCode();
+        return (info.packageName + "!" + info.applicationInfo.uid / App.PER_USER_RANGE).hashCode();
     }
 
     @Override
@@ -442,115 +497,110 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         return showList.size();
     }
 
-    public void onDestroy() {
-        loadAppListHandler.removeCallbacksAndMessages(null);
-        handlerThread.quit();
+    public void refresh() {
+        refresh(false);
     }
 
-    public void refresh() {
-        synchronized (this) {
-            if (refreshing) {
-                return;
-            }
-            refreshing = true;
-        }
-        loadAppListHandler.removeCallbacksAndMessages(null);
-        boolean force = fragment.binding.swipeRefreshLayout.isRefreshing();
-        if (!force) fragment.binding.progress.setIndeterminate(true);
+    public void refresh(boolean force) {
+        setLoaded(null, false);
         enabled = moduleUtil.isModuleEnabled(module.packageName);
-        fragment.binding.masterSwitch.setOnCheckedChangeListener(null);
-        fragment.binding.masterSwitch.setChecked(enabled);
-        fragment.binding.masterSwitch.setOnCheckedChangeListener(switchBarOnCheckedChangeListener);
-        loadAppListHandler.post(() -> {
+        fragment.runAsync(() -> {
             List<PackageInfo> appList = AppHelper.getAppList(force);
-            checkedList.clear();
-            recommendedList.clear();
-            var tmpList = new ArrayList<AppInfo>();
-            checkedList.addAll(ConfigManager.getModuleScope(module.packageName));
-            HashSet<ApplicationWithEquals> installedList = new HashSet<>();
+            denyList = AppHelper.getDenyList(force);
+            var tmpRecList = new HashSet<ApplicationWithEquals>();
+            var tmpChkList = new HashSet<>(ConfigManager.getModuleScope(module.packageName));
+            final var tmpList = new ArrayList<AppInfo>();
+            final HashSet<ApplicationWithEquals> installedList = new HashSet<>();
             List<String> scopeList = module.getScopeList();
-            boolean emptyCheckedList = checkedList.isEmpty();
-            for (PackageInfo info : appList) {
-                int userId = info.applicationInfo.uid / 100000;
+            boolean emptyCheckedList = tmpChkList.isEmpty();
+            appList.parallelStream().forEach(info -> {
+                int userId = info.applicationInfo.uid / App.PER_USER_RANGE;
                 String packageName = info.packageName;
-                if (packageName.equals("android") && userId != 0 ||
+                if (packageName.equals("system") && userId != 0 ||
                         packageName.equals(module.packageName) ||
                         packageName.equals(BuildConfig.APPLICATION_ID)) {
-                    continue;
+                    return;
                 }
 
                 ApplicationWithEquals application = new ApplicationWithEquals(packageName, userId);
 
-                installedList.add(application);
+                synchronized (installedList) {
+                    installedList.add(application);
+                }
 
                 if (userId != module.userId) {
-                    continue;
+                    return;
                 }
 
                 if (scopeList != null && scopeList.contains(packageName)) {
-                    recommendedList.add(application);
-                    if (emptyCheckedList) {
-                        checkedList.add(application);
+                    synchronized (tmpRecList) {
+                        tmpRecList.add(application);
                     }
-                } else if (shouldHideApp(info, application)) {
-                    continue;
+                    if (emptyCheckedList) {
+                        synchronized (tmpChkList) {
+                            tmpChkList.add(application);
+                        }
+                    }
+                } else if (shouldHideApp(info, application, tmpChkList)) {
+                    return;
                 }
 
                 AppInfo appInfo = new AppInfo();
                 appInfo.packageInfo = info;
-                appInfo.label = info.applicationInfo.loadLabel(pm);
+                appInfo.label = AppHelper.getAppLabel(info, pm);
                 appInfo.application = application;
                 appInfo.packageName = info.packageName;
                 appInfo.applicationInfo = info.applicationInfo;
-                tmpList.add(appInfo);
-            }
-            checkedList.retainAll(installedList);
-            if (emptyCheckedList) {
-                ConfigManager.setModuleScope(module.packageName, checkedList);
-            }
-            sortApps(tmpList);
-            searchList.clear();
-            searchList.addAll(tmpList);
-            synchronized (dataReadyRunnable) {
-                synchronized (this) {
-                    refreshing = false;
+                synchronized (tmpList) {
+                    tmpList.add(appInfo);
                 }
-                activity.runOnUiThread(dataReadyRunnable);
-                try {
-                    dataReadyRunnable.wait();
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
+            });
+            tmpChkList.retainAll(installedList);
+            checkedList = tmpChkList;
+            recommendedList = tmpRecList;
+            searchList = tmpList.parallelStream().sorted(this::sortApps).collect(Collectors.toList());
+
+            String queryStr = fragment.searchView != null ? fragment.searchView.getQuery().toString() : "";
+
+            fragment.runOnUiThread(() -> getFilter().filter(queryStr));
         });
     }
 
     protected void onCheckedChange(CompoundButton buttonView, boolean isChecked, AppInfo appInfo) {
+        var tmpChkList = new HashSet<>(checkedList);
         if (isChecked) {
-            checkedList.add(appInfo.application);
+            tmpChkList.add(appInfo.application);
         } else {
-            checkedList.remove(appInfo.application);
+            tmpChkList.remove(appInfo.application);
         }
-        if (!ConfigManager.setModuleScope(module.packageName, checkedList)) {
-            Snackbar.make(fragment.binding.snackbar, R.string.failed_to_save_scope_list, Snackbar.LENGTH_SHORT).show();
+        if (!ConfigManager.setModuleScope(module.packageName, module.legacy, tmpChkList)) {
+            fragment.showHint(R.string.failed_to_save_scope_list, true);
             if (!isChecked) {
-                checkedList.add(appInfo.application);
+                tmpChkList.add(appInfo.application);
             } else {
-                checkedList.remove(appInfo.application);
+                tmpChkList.remove(appInfo.application);
             }
             buttonView.setChecked(!isChecked);
-        } else if (appInfo.packageName.equals("android")) {
-            Snackbar.make(fragment.binding.snackbar, R.string.reboot_required, Snackbar.LENGTH_SHORT)
-                    .setAction(R.string.reboot, v -> ConfigManager.reboot(false))
-                    .show();
+        } else if (appInfo.packageName.equals("system")) {
+            fragment.showHint(R.string.reboot_required, true, R.string.reboot, v -> ConfigManager.reboot());
+        } else if (denyList.contains(appInfo.packageName)) {
+            fragment.showHint(activity.getString(R.string.deny_list, appInfo.label), true);
         }
+        checkedList = tmpChkList;
     }
 
-    static class ViewHolder extends RecyclerView.ViewHolder {
+    @Override
+    public boolean isLoaded() {
+        return isLoaded;
+    }
+
+    public static class ViewHolder extends RecyclerView.ViewHolder {
         ConstraintLayout root;
         ImageView appIcon;
         TextView appName;
-        TextView appDescription;
+        TextView appPackageName;
+        TextView appVersionName;
+        TextView hint;
         MaterialCheckBox checkbox;
 
         ViewHolder(ItemModuleBinding binding) {
@@ -558,8 +608,10 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
             root = binding.itemRoot;
             appIcon = binding.appIcon;
             appName = binding.appName;
-            appDescription = binding.description;
+            appPackageName = binding.appPackageName;
+            appVersionName = binding.appVersionName;
             checkbox = binding.checkbox;
+            hint = binding.hint;
             checkbox.setVisibility(View.VISIBLE);
         }
     }
@@ -574,15 +626,11 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         protected FilterResults performFiltering(CharSequence constraint) {
             FilterResults filterResults = new FilterResults();
             List<AppInfo> filtered = new ArrayList<>();
-            if (constraint.toString().isEmpty()) {
-                filtered.addAll(searchList);
-            } else {
-                String filter = constraint.toString().toLowerCase();
-                for (AppInfo info : searchList) {
-                    if (lowercaseContains(info.label.toString(), filter)
-                            || lowercaseContains(info.packageName, filter)) {
-                        filtered.add(info);
-                    }
+            String filter = constraint.toString().toLowerCase();
+            for (AppInfo info : searchList) {
+                if (lowercaseContains(info.label.toString(), filter)
+                        || lowercaseContains(info.packageName, filter)) {
+                    filtered.add(info);
                 }
             }
             filterResults.values = filtered;
@@ -592,10 +640,8 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
 
         @Override
         protected void publishResults(CharSequence constraint, FilterResults results) {
-            showList.clear();
             //noinspection unchecked
-            showList.addAll((Collection<AppInfo>) results.values);
-            notifyDataSetChanged();
+            setLoaded((List<AppInfo>) results.values, true);
         }
     }
 
@@ -603,13 +649,13 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
         return new SearchView.OnQueryTextListener() {
             @Override
             public boolean onQueryTextSubmit(String query) {
-                refresh();
+                getFilter().filter(query);
                 return true;
             }
 
             @Override
-            public boolean onQueryTextChange(String newText) {
-                refresh();
+            public boolean onQueryTextChange(String query) {
+                getFilter().filter(query);
                 return true;
             }
         };
@@ -617,25 +663,22 @@ public class ScopeAdapter extends RecyclerView.Adapter<ScopeAdapter.ViewHolder> 
 
     public void onBackPressed() {
         fragment.searchView.clearFocus();
-        if (!refreshing && fragment.binding.masterSwitch.isChecked() && checkedList.isEmpty()) {
-            MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(activity);
+        if (isLoaded && enabled && checkedList.isEmpty()) {
+            var builder = new BlurBehindDialogBuilder(activity, R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons);
             builder.setMessage(!recommendedList.isEmpty() ? R.string.no_scope_selected_has_recommended : R.string.no_scope_selected);
             if (!recommendedList.isEmpty()) {
-                builder.setPositiveButton(android.R.string.ok, (dialog, which) -> {
-                    checkRecommended();
-                    notifyDataSetChanged();
-                });
+                builder.setPositiveButton(android.R.string.ok, (dialog, which) -> checkRecommended());
             } else {
                 builder.setPositiveButton(android.R.string.cancel, null);
             }
             builder.setNegativeButton(!recommendedList.isEmpty() ? android.R.string.cancel : android.R.string.ok, (dialog, which) -> {
                 moduleUtil.setModuleEnabled(module.packageName, false);
                 Toast.makeText(activity, activity.getString(R.string.module_disabled_no_selection, module.getAppName()), Toast.LENGTH_LONG).show();
-                fragment.getNavController().navigateUp();
+                fragment.navigateUp();
             });
             builder.show();
         } else {
-            fragment.getNavController().navigateUp();
+            fragment.navigateUp();
         }
     }
 
