@@ -20,54 +20,59 @@
 
 package org.lsposed.manager;
 
+import android.app.ActivityManager;
 import android.app.Application;
 import android.content.BroadcastReceiver;
+import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.os.Build;
+import android.os.Environment;
+import android.os.Handler;
+import android.os.Looper;
 import android.os.Process;
+import android.provider.MediaStore;
+import android.provider.Settings;
 import android.system.Os;
 import android.text.TextUtils;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatDelegate;
 import androidx.preference.PreferenceManager;
 
-import com.google.gson.JsonParser;
-
 import org.lsposed.hiddenapibypass.HiddenApiBypass;
+import org.lsposed.manager.adapters.AppHelper;
+import org.lsposed.manager.receivers.LSPManagerServiceHolder;
 import org.lsposed.manager.repo.RepoLoader;
-import org.lsposed.manager.ui.activity.CrashReportActivity;
-import org.lsposed.manager.util.DoHDNS;
+import org.lsposed.manager.util.CloudflareDNS;
 import org.lsposed.manager.util.ModuleUtil;
+import org.lsposed.manager.util.Telemetry;
 import org.lsposed.manager.util.ThemeUtil;
+import org.lsposed.manager.util.UpdateUtil;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
-import java.time.Instant;
-import java.time.ZoneOffset;
+import java.time.OffsetDateTime;
+import java.util.HashMap;
 import java.util.Locale;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.FutureTask;
 
 import okhttp3.Cache;
-import okhttp3.Call;
-import okhttp3.Callback;
 import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
 import okhttp3.logging.HttpLoggingInterceptor;
-import rikka.material.app.DayNightDelegate;
+import rikka.core.os.FileUtils;
 import rikka.material.app.LocaleDelegate;
 
 public class App extends Application {
+    public static final int PER_USER_RANGE = 100000;
     public static final FutureTask<String> HTML_TEMPLATE = new FutureTask<>(() -> readWebviewHTML("template.html"));
     public static final FutureTask<String> HTML_TEMPLATE_DARK = new FutureTask<>(() -> readWebviewHTML("template_dark.html"));
 
@@ -75,31 +80,45 @@ public class App extends Application {
         try {
             var input = App.getInstance().getAssets().open("webview/" + name);
             var result = new ByteArrayOutputStream(1024);
-            var buffer = new byte[1024];
-            for (int length; (length = input.read(buffer)) != -1; ) {
-                result.write(buffer, 0, length);
-            }
+            FileUtils.copy(input, result);
             return result.toString(StandardCharsets.UTF_8.name());
         } catch (IOException e) {
             Log.e(App.TAG, "read webview HTML", e);
-            return "<html><body>@body@</body></html>";
+            return "<html dir\"@dir@\"><body>@body@</body></html>";
         }
     }
 
     static {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-            // TODO: set specific class name
             HiddenApiBypass.addHiddenApiExemptions("");
         }
+        Looper.myQueue().addIdleHandler(() -> {
+            if (App.getInstance() == null || App.getExecutorService() == null) return true;
+            App.getExecutorService().submit(() -> {
+                var list = AppHelper.getAppList(false);
+                var pm = App.getInstance().getPackageManager();
+                list.parallelStream().forEach(i -> AppHelper.getAppLabel(i, pm));
+                AppHelper.getDenyList(false);
+                ModuleUtil.getInstance();
+                RepoLoader.getInstance();
+            });
+            App.getExecutorService().submit(HTML_TEMPLATE);
+            App.getExecutorService().submit(HTML_TEMPLATE_DARK);
+            return false;
+        });
     }
 
     public static final String TAG = "LSPosedManager";
+    private static final String ACTION_USER_ADDED = "android.intent.action.USER_ADDED";
+    private static final String ACTION_USER_REMOVED = "android.intent.action.USER_REMOVED";
+    private static final String ACTION_USER_INFO_CHANGED = "android.intent.action.USER_INFO_CHANGED";
+    private static final String EXTRA_REMOVED_FOR_ALL_USERS = "android.intent.extra.REMOVED_FOR_ALL_USERS";
     private static App instance = null;
     private static OkHttpClient okHttpClient;
     private static Cache okHttpCache;
-    private static boolean parasiticShown = false;
     private SharedPreferences pref;
-    private ExecutorService executorService;
+    private static final ExecutorService executorService = Executors.newCachedThreadPool();
+    private static final Handler MainHandler = new Handler(Looper.getMainLooper());
 
     public static App getInstance() {
         return instance;
@@ -110,168 +129,155 @@ public class App extends Application {
     }
 
     public static ExecutorService getExecutorService() {
-        return instance.executorService;
+        return executorService;
     }
 
-    public static boolean isParasitic() {
-        return !Process.isApplicationUid(Process.myUid());
+    public static final boolean isParasitic = !Process.isApplicationUid(Process.myUid());
+
+    public static Handler getMainHandler() {
+        return MainHandler;
     }
 
-    public static boolean isParasiticShown() {
-        return parasiticShown;
-    }
-
-    public static void setParasiticShown(boolean parasiticShown) {
-        App.parasiticShown = parasiticShown;
+    @Override
+    protected void attachBaseContext(Context base) {
+        super.attachBaseContext(base);
+        Telemetry.start(this);
+        var map = new HashMap<String, String>(1);
+        map.put("isParasitic", String.valueOf(isParasitic));
+        Telemetry.trackEvent("App start", map);
+        var am = getSystemService(ActivityManager.class);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            map.clear();
+            var reasons = am.getHistoricalProcessExitReasons(null, 0, 1);
+            if (reasons.size() == 1) {
+                map.put("description", reasons.get(0).getDescription());
+                map.put("importance", String.valueOf(reasons.get(0).getImportance()));
+                map.put("process", reasons.get(0).getProcessName());
+                map.put("reason", String.valueOf(reasons.get(0).getReason()));
+                map.put("status", String.valueOf(reasons.get(0).getStatus()));
+                Telemetry.trackEvent("Last exit reasons", map);
+            }
+        }
     }
 
     private void setCrashReport() {
+        var handler = Thread.getDefaultUncaughtExceptionHandler();
         Thread.setDefaultUncaughtExceptionHandler((thread, throwable) -> {
-
-            StringWriter sw = new StringWriter();
-            PrintWriter pw = new PrintWriter(sw);
-            throwable.printStackTrace(pw);
-            String stackTraceString = sw.toString();
-
-            //Reduce data to 128KB so we don't get a TransactionTooLargeException when sending the intent.
-            //The limit is 1MB on Android but some devices seem to have it lower.
-            //See: http://developer.android.com/reference/android/os/TransactionTooLargeException.html
-            //And: http://stackoverflow.com/questions/11451393/what-to-do-on-transactiontoolargeexception#comment46697371_12809171
-            if (stackTraceString.length() > 131071) {
-                String disclaimer = " [stack trace too large]";
-                stackTraceString = stackTraceString.substring(0, 131071 - disclaimer.length()) + disclaimer;
+            var time = OffsetDateTime.now();
+            var dir = new File(getCacheDir(), "crash");
+            //noinspection ResultOfMethodCallIgnored
+            dir.mkdir();
+            var file = new File(dir, time.toEpochSecond() + ".log");
+            try (var pw = new PrintWriter(file)) {
+                pw.println(BuildConfig.VERSION_NAME + " (" + BuildConfig.VERSION_CODE + ")");
+                pw.println(time);
+                pw.println("pid: " + Os.getpid() + " uid: " + Os.getuid());
+                throwable.printStackTrace(pw);
+            } catch (IOException ignored) {
             }
-            Intent intent = new Intent(App.this, CrashReportActivity.class);
-            intent.putExtra(BuildConfig.APPLICATION_ID + ".EXTRA_STACK_TRACE", stackTraceString);
-            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-            App.this.startActivity(intent);
-            System.exit(10);
-            Process.killProcess(Os.getpid());
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+                var table = MediaStore.Downloads.getContentUri(MediaStore.VOLUME_EXTERNAL_PRIMARY);
+                var values = new ContentValues();
+                values.put(MediaStore.Downloads.DISPLAY_NAME, "LSPosed_crash_report" + time.toEpochSecond() + ".zip");
+                values.put(MediaStore.Downloads.RELATIVE_PATH, Environment.DIRECTORY_DOCUMENTS);
+                var cr = getContentResolver();
+                var uri = cr.insert(table, values);
+                if (uri == null) return;
+                try (var zipFd = cr.openFileDescriptor(uri, "wt")) {
+                    LSPManagerServiceHolder.getService().getLogs(zipFd);
+                } catch (Exception ignored) {
+                    cr.delete(uri, null, null);
+                }
+            }
+            if (handler != null) {
+                handler.uncaughtException(thread, throwable);
+            }
         });
     }
 
     @Override
     public void onCreate() {
         super.onCreate();
-        if (!BuildConfig.DEBUG && !isParasitic()) {
-            setCrashReport();
-        }
-
         instance = this;
 
-        executorService = Executors.newCachedThreadPool();
-
+        setCrashReport();
         pref = PreferenceManager.getDefaultSharedPreferences(this);
-        if ("CN".equals(Locale.getDefault().getCountry())) {
-            if (!pref.contains("doh")) {
+        if (!pref.contains("doh")) {
+            var name = "private_dns_mode";
+            if ("hostname".equals(Settings.Global.getString(getContentResolver(), name))) {
+                pref.edit().putBoolean("doh", false).apply();
+            } else {
                 pref.edit().putBoolean("doh", true).apply();
             }
         }
-        DayNightDelegate.setApplicationContext(this);
-        DayNightDelegate.setDefaultNightMode(ThemeUtil.getDarkTheme());
+        AppCompatDelegate.setDefaultNightMode(ThemeUtil.getDarkTheme());
         LocaleDelegate.setDefaultLocale(getLocale());
+        var res = getResources();
+        var config = res.getConfiguration();
+        config.setLocale(LocaleDelegate.getDefaultLocale());
+        //noinspection deprecation
+        res.updateConfiguration(config, res.getDisplayMetrics());
 
+        IntentFilter intentFilter = new IntentFilter();
+        intentFilter.addAction("org.lsposed.manager.NOTIFICATION");
         registerReceiver(new BroadcastReceiver() {
             @Override
-            public void onReceive(Context context, Intent intent) {
-                int userId = intent.getIntExtra(Intent.EXTRA_USER, 0);
-                String packageName = intent.getStringExtra("android.intent.extra.PACKAGES");
-                boolean packageFullyRemoved = intent.getBooleanExtra(Intent.ACTION_PACKAGE_FULLY_REMOVED, false);
-                if (packageName != null) {
-                    ModuleUtil.getInstance().reloadSingleModule(packageName, userId, packageFullyRemoved);
+            public void onReceive(Context context, Intent inIntent) {
+                var intent = (Intent) inIntent.getParcelableExtra(Intent.EXTRA_INTENT);
+                Log.d(TAG, "onReceive: " + intent);
+                switch (intent.getAction()) {
+                    case Intent.ACTION_PACKAGE_ADDED, Intent.ACTION_PACKAGE_CHANGED, Intent.ACTION_PACKAGE_FULLY_REMOVED, Intent.ACTION_UID_REMOVED -> {
+                        var userId = intent.getIntExtra(Intent.EXTRA_USER, 0);
+                        var packageName = intent.getStringExtra("android.intent.extra.PACKAGES");
+                        var packageRemovedForAllUsers = intent.getBooleanExtra(EXTRA_REMOVED_FOR_ALL_USERS, false);
+                        var isXposedModule = intent.getBooleanExtra("isXposedModule", false);
+                        if (packageName != null) {
+                            if (isXposedModule)
+                                ModuleUtil.getInstance().reloadSingleModule(packageName, userId, packageRemovedForAllUsers);
+                            else
+                                App.getExecutorService().submit(() -> AppHelper.getAppList(true));
+                        }
+                    }
+                    case ACTION_USER_ADDED, ACTION_USER_REMOVED, ACTION_USER_INFO_CHANGED -> App.getExecutorService().submit(() -> ModuleUtil.getInstance().reloadInstalledModules());
                 }
             }
-        }, new IntentFilter(Intent.ACTION_PACKAGE_CHANGED));
+        }, intentFilter, Context.RECEIVER_NOT_EXPORTED);
 
-        loadRemoteVersion();
-        RepoLoader.getInstance().loadRemoteData();
-
-        executorService.submit(HTML_TEMPLATE);
-        executorService.submit(HTML_TEMPLATE_DARK);
+        UpdateUtil.loadRemoteVersion();
     }
 
     @NonNull
     public static OkHttpClient getOkHttpClient() {
-        if (okHttpClient == null) {
-            OkHttpClient.Builder builder = new OkHttpClient.Builder().cache(getOkHttpCache());
-            builder.addInterceptor(chain -> {
-                var request = chain.request().newBuilder();
-                request.header("User-Agent", TAG);
-                return chain.proceed(request.build());
-            });
-            HttpLoggingInterceptor log = new HttpLoggingInterceptor();
+        if (okHttpClient != null) return okHttpClient;
+        var builder = new OkHttpClient.Builder()
+            .cache(getOkHttpCache())
+            .dns(new CloudflareDNS());
+        if (BuildConfig.DEBUG) {
+            var log = new HttpLoggingInterceptor();
             log.setLevel(HttpLoggingInterceptor.Level.HEADERS);
-            if (BuildConfig.DEBUG) builder.addInterceptor(log);
-            okHttpClient = builder.dns(new DoHDNS(builder.build())).build();
+            builder.addInterceptor(log);
         }
+        okHttpClient = builder.build();
         return okHttpClient;
     }
 
     @NonNull
-    private static Cache getOkHttpCache() {
-        if (okHttpCache == null) {
-            okHttpCache = new Cache(new File(App.getInstance().getCacheDir(), "http_cache"), 50L * 1024L * 1024L);
-        }
+    public static Cache getOkHttpCache() {
+        if (okHttpCache != null) return okHttpCache;
+        long size50MiB = 50 * 1024 * 1024;
+        okHttpCache = new Cache(new File(instance.getCacheDir(), "http_cache"), size50MiB);
         return okHttpCache;
     }
 
-    private void loadRemoteVersion() {
-        var request = new Request.Builder()
-                .url("https://api.github.com/repos/LSPosed/LSPosed/releases/latest")
-                .addHeader("Accept", "application/vnd.github.v3+json")
-                .build();
-        var callback = new Callback() {
-            @Override
-            public void onResponse(@NonNull Call call, @NonNull Response response) {
-                if (!response.isSuccessful()) return;
-                var body = response.body();
-                if (body == null) return;
-                try {
-                    var info = JsonParser.parseReader(body.charStream()).getAsJsonObject();
-                    var name = info.getAsJsonArray("assets").get(0).getAsJsonObject().get("name").getAsString();
-                    var code = Integer.parseInt(name.split("-", 4)[2]);
-                    var now = Instant.now().getEpochSecond();
-                    pref.edit()
-                            .putInt("latest_version", code)
-                            .putLong("latest_check", now)
-                            .putBoolean("checked", true)
-                            .apply();
-                } catch (Throwable t) {
-                    Log.e(App.TAG, t.getMessage(), t);
-                }
-            }
-
-            @Override
-            public void onFailure(@NonNull Call call, @NonNull IOException e) {
-                Log.e(App.TAG, "loadRemoteVersion: " + e.getMessage());
-                if (pref.getBoolean("checked", false)) return;
-                pref.edit().putBoolean("checked", true).apply();
-            }
-        };
-        getOkHttpClient().newCall(request).enqueue(callback);
-    }
-
-    public static boolean needUpdate() {
-        var pref = getPreferences();
-        if (!pref.getBoolean("checked", false)) return false;
-        var now = Instant.now();
-        var buildTime = Instant.ofEpochSecond(BuildConfig.BUILD_TIME);
-        var check = pref.getLong("latest_check", 0);
-        if (check > 0) {
-            var checkTime = Instant.ofEpochSecond(check);
-            if (checkTime.atOffset(ZoneOffset.UTC).plusDays(30).toInstant().isBefore(now))
-                return true;
-            var code = pref.getInt("latest_version", 0);
-            return code > BuildConfig.VERSION_CODE;
+    public static Locale getLocale(String tag) {
+        if (TextUtils.isEmpty(tag) || "SYSTEM".equals(tag)) {
+            return LocaleDelegate.getSystemLocale();
         }
-        return buildTime.atOffset(ZoneOffset.UTC).plusDays(30).toInstant().isBefore(now);
+        return Locale.forLanguageTag(tag);
     }
 
     public static Locale getLocale() {
         String tag = getPreferences().getString("language", null);
-        if (TextUtils.isEmpty(tag) || "SYSTEM".equals(tag)) {
-            return Locale.getDefault();
-        }
-        return Locale.forLanguageTag(tag);
+        return getLocale(tag);
     }
 }
