@@ -20,28 +20,32 @@
 
 package org.lsposed.manager.repo;
 
+import android.content.res.Resources;
 import android.util.Log;
-import android.util.Pair;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.google.gson.Gson;
 
 import org.lsposed.manager.App;
+import org.lsposed.manager.R;
 import org.lsposed.manager.repo.model.OnlineModule;
+import org.lsposed.manager.repo.model.Release;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CopyOnWriteArrayList;
 
 import okhttp3.Call;
 import okhttp3.Callback;
@@ -52,14 +56,33 @@ import okhttp3.ResponseBody;
 public class RepoLoader {
     private static RepoLoader instance = null;
     private Map<String, OnlineModule> onlineModules = new HashMap<>();
-    private final Map<String, Pair<Integer, String>> latestVersion = new ConcurrentHashMap<>();
+    private Map<String, ModuleVersion> latestVersion = new ConcurrentHashMap<>();
+
+    public static class ModuleVersion {
+        public String versionName;
+        public long versionCode;
+
+        private ModuleVersion(long versionCode, String versionName) {
+            this.versionName = versionName;
+            this.versionCode = versionCode;
+        }
+
+        public boolean upgradable(long versionCode, String versionName) {
+            return this.versionCode > versionCode || (this.versionCode == versionCode && !versionName.replace(' ', '_').equals(this.versionName));
+        }
+
+    }
+
     private final Path repoFile = Paths.get(App.getInstance().getFilesDir().getAbsolutePath(), "repo.json");
-    private final List<Listener> listeners = new CopyOnWriteArrayList<>();
-    private boolean isLoading = false;
+    private final Set<RepoListener> listeners = ConcurrentHashMap.newKeySet();
     private boolean repoLoaded = false;
     private static final String originRepoUrl = "https://modules.lsposed.org/";
-    private static final String backupRepoUrl = "https://cdn.jsdelivr.net/gh/Xposed-Modules-Repo/modules@gh-pages/";
+    private static final String backupRepoUrl = "https://modules-blogcdn.lsposed.org/";
+
+    private static final String secondBackupRepoUrl = "https://modules-cloudflare.lsposed.org/";
     private static String repoUrl = originRepoUrl;
+    private final Resources resources = App.getInstance().getResources();
+    private final String[] channels = resources.getStringArray(R.array.update_channel_values);
 
     public boolean isRepoLoaded() {
         return repoLoaded;
@@ -68,105 +91,175 @@ public class RepoLoader {
     public static synchronized RepoLoader getInstance() {
         if (instance == null) {
             instance = new RepoLoader();
-            instance.loadRemoteData();
+            App.getExecutorService().submit(() -> instance.loadLocalData(true));
         }
         return instance;
     }
 
-    public void loadRemoteData() {
-        synchronized (this) {
-            if (isLoading) {
-                return;
-            }
-            isLoading = true;
-        }
-        App.getOkHttpClient().newCall(new Request.Builder()
-                .url(repoUrl + "modules.json")
-                .build()).enqueue(new Callback() {
-            @Override
-            public void onFailure(@NonNull Call call, @NonNull IOException e) {
-                Log.e(App.TAG, call.request().url() + e.getMessage());
-                for (Listener listener : listeners) {
-                    listener.onThrowable(e);
-                }
-                synchronized (this) {
-                    isLoading = false;
-                    if (!repoUrl.equals(backupRepoUrl)) {
-                        repoUrl = backupRepoUrl;
-                        loadRemoteData();
-                    }
-                }
-            }
+    synchronized public void loadRemoteData() {
+        repoLoaded = false;
+        try {
+            try (var response = App.getOkHttpClient().newCall(new Request.Builder().url(repoUrl + "modules.json").build()).execute()) {
 
-            @Override
-            public void onResponse(@NonNull Call call, @NonNull Response response) {
                 if (response.isSuccessful()) {
                     ResponseBody body = response.body();
                     if (body != null) {
                         try {
                             String bodyString = body.string();
-                            Gson gson = new Gson();
-                            Map<String, OnlineModule> modules = new HashMap<>();
-                            OnlineModule[] repoModules = gson.fromJson(bodyString, OnlineModule[].class);
-                            Arrays.stream(repoModules).forEach(onlineModule -> modules.put(onlineModule.getName(), onlineModule));
-
-                            latestVersion.clear();
-                            for (var module : repoModules) {
-                                var release = module.getLatestRelease();
-                                if (release == null || release.isEmpty()) continue;
-                                var splits = release.split("-", 2);
-                                if (splits.length < 2) continue;
-                                int verCode;
-                                String verName;
-                                try {
-                                    verCode = Integer.parseInt(splits[0]);
-                                    verName = splits[1];
-                                } catch (NumberFormatException ignored) {
-                                    continue;
-                                }
-                                String pkgName = module.getName();
-                                latestVersion.put(pkgName, new Pair<>(verCode, verName));
-                            }
-
-                            onlineModules = modules;
                             Files.write(repoFile, bodyString.getBytes(StandardCharsets.UTF_8));
-                            for (Listener listener : listeners) {
-                                listener.repoLoaded();
-                            }
-                            synchronized (this) {
-                                repoLoaded = true;
-                            }
+                            loadLocalData(false);
                         } catch (Throwable t) {
                             Log.e(App.TAG, Log.getStackTraceString(t));
-                            for (Listener listener : listeners) {
+                            for (RepoListener listener : listeners) {
                                 listener.onThrowable(t);
                             }
                         }
                     }
                 }
-                synchronized (this) {
-                    isLoading = false;
-                }
             }
-        });
+        } catch (Throwable e) {
+            Log.e(App.TAG, "load remote data", e);
+            for (RepoListener listener : listeners) {
+                listener.onThrowable(e);
+            }
+            if (repoUrl.equals(originRepoUrl)) {
+                repoUrl = backupRepoUrl;
+                loadRemoteData();
+            } else if (repoUrl.equals(backupRepoUrl)) {
+                repoUrl = secondBackupRepoUrl;
+                loadRemoteData();
+            }
+        }
     }
 
-    public Pair<Integer, String> getModuleLatestVersion(String packageName) {
-        return latestVersion.get(packageName);
+    synchronized public void loadLocalData(boolean updateRemoteRepo) {
+        repoLoaded = false;
+        try {
+            if (Files.notExists(repoFile)) {
+                loadRemoteData();
+                updateRemoteRepo = false;
+            }
+            byte[] encoded = Files.readAllBytes(repoFile);
+            String bodyString = new String(encoded, StandardCharsets.UTF_8);
+            Gson gson = new Gson();
+            Map<String, OnlineModule> modules = new HashMap<>();
+            OnlineModule[] repoModules = gson.fromJson(bodyString, OnlineModule[].class);
+            Arrays.stream(repoModules).forEach(onlineModule -> modules.put(onlineModule.getName(), onlineModule));
+            var channel = App.getPreferences().getString("update_channel", channels[0]);
+            updateLatestVersion(repoModules, channel);
+            onlineModules = modules;
+        } catch (Throwable t) {
+            Log.e(App.TAG, Log.getStackTraceString(t));
+            for (RepoListener listener : listeners) {
+                listener.onThrowable(t);
+            }
+        } finally {
+            repoLoaded = true;
+            for (RepoListener listener : listeners) {
+                listener.onRepoLoaded();
+            }
+            if (updateRemoteRepo) loadRemoteData();
+        }
+    }
+
+    synchronized private void updateLatestVersion(OnlineModule[] onlineModules, String channel) {
+        repoLoaded = false;
+        Map<String, ModuleVersion> versions = new ConcurrentHashMap<>();
+        for (var module : onlineModules) {
+            String release = module.getLatestRelease();
+            if (channel.equals(channels[1]) && module.getLatestBetaRelease() != null && !module.getLatestBetaRelease().isEmpty()) {
+                release = module.getLatestBetaRelease();
+            } else if (channel.equals(channels[2])) {
+                if (module.getLatestSnapshotRelease() != null && !module.getLatestSnapshotRelease().isEmpty())
+                    release = module.getLatestSnapshotRelease();
+                else if (module.getLatestBetaRelease() != null && !module.getLatestBetaRelease().isEmpty())
+                    release = module.getLatestBetaRelease();
+            }
+            if (release == null || release.isEmpty()) continue;
+            var splits = release.split("-", 2);
+            if (splits.length < 2) continue;
+            long verCode;
+            String verName;
+            try {
+                verCode = Long.parseLong(splits[0]);
+                verName = splits[1];
+            } catch (NumberFormatException ignored) {
+                continue;
+            }
+            String pkgName = module.getName();
+            versions.put(pkgName, new ModuleVersion(verCode, verName));
+        }
+        latestVersion = versions;
+        repoLoaded = true;
+        for (RepoListener listener : listeners) {
+            listener.onRepoLoaded();
+        }
+    }
+
+    public void updateLatestVersion(String channel) {
+        if (repoLoaded)
+            updateLatestVersion(onlineModules.keySet().parallelStream().map(onlineModules::get).toArray(OnlineModule[]::new), channel);
+    }
+
+    @Nullable
+    public ModuleVersion getModuleLatestVersion(String packageName) {
+        return repoLoaded ? latestVersion.getOrDefault(packageName, null) : null;
+    }
+
+    @Nullable
+    public List<Release> getReleases(String packageName) {
+        var channel = App.getPreferences().getString("update_channel", channels[0]);
+        List<Release> releases = new ArrayList<>();
+        if (repoLoaded) {
+            var module = onlineModules.get(packageName);
+            if (module != null) {
+                releases = module.getReleases();
+                if (!module.releasesLoaded) {
+                    if (channel.equals(channels[1]) && !(module.getBetaReleases() != null && module.getBetaReleases().isEmpty())) {
+                        releases = module.getBetaReleases();
+                    } else if (channel.equals(channels[2]))
+                        if (!(module.getSnapshotReleases() != null && module.getSnapshotReleases().isEmpty()))
+                            releases = module.getSnapshotReleases();
+                        else if (!(module.getBetaReleases() != null && module.getBetaReleases().isEmpty()))
+                            releases = module.getBetaReleases();
+                }
+            }
+        }
+        return releases;
+    }
+
+    @Nullable
+    public String getLatestReleaseTime(String packageName, String channel) {
+        String releaseTime = null;
+        if (repoLoaded) {
+            var module = onlineModules.get(packageName);
+            if (module != null) {
+                releaseTime = module.getLatestReleaseTime();
+                if (channel.equals(channels[1]) && module.getLatestBetaReleaseTime() != null) {
+                    releaseTime = module.getLatestBetaReleaseTime();
+                } else if (channel.equals(channels[2]))
+                    if (module.getLatestSnapshotReleaseTime() != null)
+                        releaseTime = module.getLatestSnapshotReleaseTime();
+                    else if (module.getLatestBetaReleaseTime() != null)
+                        releaseTime = module.getLatestBetaReleaseTime();
+            }
+        }
+        return releaseTime;
     }
 
     public void loadRemoteReleases(String packageName) {
-        App.getOkHttpClient().newCall(new Request.Builder()
-                .url(String.format(repoUrl + "module/%s.json", packageName))
-                .build()).enqueue(new Callback() {
+        App.getOkHttpClient().newCall(new Request.Builder().url(String.format(repoUrl + "module/%s.json", packageName)).build()).enqueue(new Callback() {
             @Override
             public void onFailure(@NonNull Call call, @NonNull IOException e) {
                 Log.e(App.TAG, call.request().url() + e.getMessage());
-                if (!repoUrl.equals(backupRepoUrl)) {
+                if (repoUrl.equals(originRepoUrl)) {
                     repoUrl = backupRepoUrl;
                     loadRemoteReleases(packageName);
+                } else if (repoUrl.equals(backupRepoUrl)) {
+                    repoUrl = secondBackupRepoUrl;
+                    loadRemoteReleases(packageName);
                 } else {
-                    for (Listener listener : listeners) {
+                    for (RepoListener listener : listeners) {
                         listener.onThrowable(e);
                     }
                 }
@@ -183,12 +276,12 @@ public class RepoLoader {
                             OnlineModule module = gson.fromJson(bodyString, OnlineModule.class);
                             module.releasesLoaded = true;
                             onlineModules.replace(packageName, module);
-                            for (Listener listener : listeners) {
-                                listener.moduleReleasesLoaded(module);
+                            for (RepoListener listener : listeners) {
+                                listener.onModuleReleasesLoaded(module);
                             }
                         } catch (Throwable t) {
                             Log.e(App.TAG, Log.getStackTraceString(t));
-                            for (Listener listener : listeners) {
+                            for (RepoListener listener : listeners) {
                                 listener.onThrowable(t);
                             }
                         }
@@ -198,28 +291,29 @@ public class RepoLoader {
         });
     }
 
-    public void addListener(Listener listener) {
-        if (!listeners.contains(listener))
-            listeners.add(listener);
+    public void addListener(RepoListener listener) {
+        listeners.add(listener);
     }
 
-    public void removeListener(Listener listener) {
+    public void removeListener(RepoListener listener) {
         listeners.remove(listener);
     }
 
+    @Nullable
     public OnlineModule getOnlineModule(String packageName) {
-        return packageName == null ? null : onlineModules.get(packageName);
+        return repoLoaded && packageName != null ? onlineModules.get(packageName) : null;
     }
 
+    @Nullable
     public Collection<OnlineModule> getOnlineModules() {
-        return onlineModules.values();
+        return repoLoaded ? onlineModules.values() : null;
     }
 
-    public interface Listener {
-        default void repoLoaded() {
+    public interface RepoListener {
+        default void onRepoLoaded() {
         }
 
-        default void moduleReleasesLoaded(OnlineModule module) {
+        default void onModuleReleasesLoaded(OnlineModule module) {
         }
 
         default void onThrowable(Throwable t) {
