@@ -20,6 +20,8 @@
 
 package org.lsposed.manager.repo.model;
 
+import androidx.annotation.Nullable;
+
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
@@ -46,9 +48,30 @@ public class OnlineModule {
     @SerializedName("latestRelease")
     @Expose
     private String latestRelease;
+    @SerializedName("latestReleaseTime")
+    @Expose
+    private String latestReleaseTime;
+    @SerializedName("latestBetaRelease")
+    @Expose
+    private String latestBetaRelease;
+    @SerializedName("latestBetaReleaseTime")
+    @Expose
+    private String latestBetaReleaseTime;
+    @SerializedName("latestSnapshotRelease")
+    @Expose
+    private String latestSnapshotRelease;
+    @SerializedName("latestSnapshotReleaseTime")
+    @Expose
+    private String latestSnapshotReleaseTime;
     @SerializedName("releases")
     @Expose
     private List<Release> releases = new ArrayList<>();
+    @SerializedName("betaReleases")
+    @Expose
+    private final List<Release> betaReleases = new ArrayList<>();
+    @SerializedName("snapshotReleases")
+    @Expose
+    private final List<Release> snapshotReleases = new ArrayList<>();
     @SerializedName("readme")
     @Expose
     private String readme;
@@ -81,6 +104,7 @@ public class OnlineModule {
     private Integer stargazerCount;
     public boolean releasesLoaded = false;
 
+    @Nullable
     public String getName() {
         return name;
     }
@@ -89,6 +113,7 @@ public class OnlineModule {
         this.name = name;
     }
 
+    @Nullable
     public String getDescription() {
         return description;
     }
@@ -97,6 +122,7 @@ public class OnlineModule {
         this.description = description;
     }
 
+    @Nullable
     public String getUrl() {
         return url;
     }
@@ -105,6 +131,7 @@ public class OnlineModule {
         this.url = url;
     }
 
+    @Nullable
     public String getHomepageUrl() {
         return homepageUrl;
     }
@@ -113,6 +140,7 @@ public class OnlineModule {
         this.homepageUrl = homepageUrl;
     }
 
+    @Nullable
     public List<Collaborator> getCollaborators() {
         return collaborators;
     }
@@ -121,14 +149,21 @@ public class OnlineModule {
         this.collaborators = collaborators;
     }
 
+    @Nullable
     public List<Release> getReleases() {
         return releases;
+    }
+
+    @Nullable
+    public String getLatestReleaseTime() {
+        return latestReleaseTime;
     }
 
     public void setReleases(List<Release> releases) {
         this.releases = releases;
     }
 
+    @Nullable
     public String getReadme() {
         return readme;
     }
@@ -137,6 +172,7 @@ public class OnlineModule {
         this.readme = readme;
     }
 
+    @Nullable
     public String getReadmeHTML() {
         return readmeHTML;
     }
@@ -145,7 +181,7 @@ public class OnlineModule {
         this.readmeHTML = readmeHTML;
     }
 
-
+    @Nullable
     public String getSummary() {
         return summary;
     }
@@ -154,6 +190,7 @@ public class OnlineModule {
         this.summary = summary;
     }
 
+    @Nullable
     public List<String> getScope() {
         return scope;
     }
@@ -162,6 +199,7 @@ public class OnlineModule {
         this.scope = scope;
     }
 
+    @Nullable
     public String getSourceUrl() {
         return sourceUrl;
     }
@@ -178,6 +216,7 @@ public class OnlineModule {
         this.hide = hide;
     }
 
+    @Nullable
     public List<Object> getAdditionalAuthors() {
         return additionalAuthors;
     }
@@ -186,6 +225,7 @@ public class OnlineModule {
         this.additionalAuthors = additionalAuthors;
     }
 
+    @Nullable
     public String getUpdatedAt() {
         return updatedAt;
     }
@@ -194,6 +234,7 @@ public class OnlineModule {
         this.updatedAt = updatedAt;
     }
 
+    @Nullable
     public String getCreatedAt() {
         return createdAt;
     }
@@ -202,6 +243,7 @@ public class OnlineModule {
         this.createdAt = createdAt;
     }
 
+    @Nullable
     public Integer getStargazerCount() {
         return stargazerCount;
     }
@@ -210,11 +252,42 @@ public class OnlineModule {
         this.stargazerCount = stargazerCount;
     }
 
+    @Nullable
     public String getLatestRelease() {
         return latestRelease;
     }
 
     public void setLatestRelease(String latestRelease) {
         this.latestRelease = latestRelease;
+    }
+
+    @Nullable
+    public String getLatestBetaRelease() {
+        return latestBetaRelease;
+    }
+
+    @Nullable
+    public String getLatestBetaReleaseTime() {
+        return latestBetaReleaseTime;
+    }
+
+    @Nullable
+    public String getLatestSnapshotRelease() {
+        return latestSnapshotRelease;
+    }
+
+    @Nullable
+    public String getLatestSnapshotReleaseTime() {
+        return latestSnapshotReleaseTime;
+    }
+
+    @Nullable
+    public List<Release> getBetaReleases() {
+        return betaReleases;
+    }
+
+    @Nullable
+    public List<Release> getSnapshotReleases() {
+        return snapshotReleases;
     }
 }
