@@ -19,14 +19,23 @@
 
 package org.lsposed.manager.ui.fragment;
 
+import android.annotation.SuppressLint;
+import android.app.Activity;
+import android.app.Dialog;
+import android.content.res.Resources;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.text.Spannable;
 import android.text.SpannableStringBuilder;
 import android.text.TextUtils;
+import android.text.format.Formatter;
 import android.text.style.ClickableSpan;
+import android.text.style.ForegroundColorSpan;
+import android.text.style.RelativeSizeSpan;
 import android.util.Log;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
@@ -35,18 +44,22 @@ import android.webkit.WebResourceResponse;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+import android.widget.ArrayAdapter;
+import android.widget.ScrollView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.core.view.MenuProvider;
+import androidx.fragment.app.DialogFragment;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.viewpager2.widget.ViewPager2;
+import androidx.viewpager2.adapter.FragmentStateAdapter;
 
 import com.google.android.material.button.MaterialButton;
-import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.android.material.progressindicator.CircularProgressIndicator;
-import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.tabs.TabLayout;
 import com.google.android.material.tabs.TabLayoutMediator;
 
@@ -63,31 +76,38 @@ import org.lsposed.manager.repo.model.Collaborator;
 import org.lsposed.manager.repo.model.OnlineModule;
 import org.lsposed.manager.repo.model.Release;
 import org.lsposed.manager.repo.model.ReleaseAsset;
+import org.lsposed.manager.ui.dialog.BlurBehindDialogBuilder;
+import org.lsposed.manager.ui.widget.EmptyStateRecyclerView;
 import org.lsposed.manager.ui.widget.LinkifyTextView;
 import org.lsposed.manager.util.NavUtil;
+import org.lsposed.manager.util.SimpleStatefulAdaptor;
 import org.lsposed.manager.util.chrome.CustomTabsURLSpan;
 
 import java.io.ByteArrayInputStream;
 import java.nio.charset.StandardCharsets;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.ListIterator;
+import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import okhttp3.Headers;
 import okhttp3.Request;
 import okhttp3.Response;
 import rikka.core.util.ResourceUtils;
-import rikka.insets.WindowInsetsHelperKt;
+import rikka.material.app.LocaleDelegate;
 import rikka.recyclerview.RecyclerViewKt;
-import rikka.widget.borderview.BorderNestedScrollView;
-import rikka.widget.borderview.BorderRecyclerView;
 import rikka.widget.borderview.BorderView;
 
-public class RepoItemFragment extends BaseFragment implements RepoLoader.Listener {
+public class RepoItemFragment extends BaseFragment implements RepoLoader.RepoListener, MenuProvider {
     FragmentPagerBinding binding;
-    private OnlineModule module;
+    OnlineModule module;
     private ReleaseAdapter releaseAdapter;
+    private InformationAdapter informationAdapter;
 
     @Nullable
     @Override
@@ -96,21 +116,11 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
         if (module == null) return binding.getRoot();
         String modulePackageName = module.getName();
         String moduleName = module.getDescription();
-        binding.getRoot().bringChildToFront(binding.appBar);
-        setupToolbar(binding.toolbar, moduleName, R.menu.menu_repo_item);
         binding.appBar.setLiftable(true);
+        setupToolbar(binding.toolbar, binding.clickView, moduleName, R.menu.menu_repo_item);
+        binding.clickView.setTooltipText(moduleName);
         binding.toolbar.setSubtitle(modulePackageName);
-        binding.viewPager.setAdapter(new PagerAdapter());
-        binding.viewPager.registerOnPageChangeCallback(new ViewPager2.OnPageChangeCallback() {
-            @Override
-            public void onPageSelected(int position) {
-                BorderView borderView = binding.viewPager.findViewWithTag(position);
-
-                if (borderView != null) {
-                    binding.appBar.setLifted(!borderView.getBorderViewDelegate().isShowingTopBorder());
-                }
-            }
-        });
+        binding.viewPager.setAdapter(new PagerAdapter(this));
         int[] titles = new int[]{R.string.module_readme, R.string.module_releases, R.string.module_information};
         new TabLayoutMediator(binding.tabLayout, binding.viewPager, (tab, position) -> tab.setText(titles[position])).attach();
 
@@ -122,13 +132,11 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
                 binding.tabLayout.setTabGravity(TabLayout.GRAVITY_FILL);
             }
         });
-
+        binding.toolbar.setOnClickListener(v -> binding.appBar.setExpanded(true, true));
+        releaseAdapter = new ReleaseAdapter();
+        informationAdapter = new InformationAdapter();
+        RepoLoader.getInstance().addListener(this);
         return binding.getRoot();
-    }
-
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
     }
 
     @Override
@@ -138,17 +146,19 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
 
         String modulePackageName = getArguments() == null ? null : getArguments().getString("modulePackageName");
         module = RepoLoader.getInstance().getOnlineModule(modulePackageName);
-        if (module == null)
-            getNavController().navigate(R.id.action_repo_item_fragment_to_repo_fragment);
+        if (module == null) {
+            if (!safeNavigate(R.id.action_repo_item_fragment_to_repo_fragment)) {
+                safeNavigate(R.id.repo_nav);
+            }
+        }
     }
 
-    private void renderGithubMarkdown(WebView view, String text) {
+    private void renderGithubMarkdown(WebView view, @Nullable String text) {
         try {
             view.setBackgroundColor(Color.TRANSPARENT);
             var setting = view.getSettings();
             setting.setOffscreenPreRaster(true);
             setting.setDomStorageEnabled(true);
-            setting.setAppCacheEnabled(true);
             setting.setMixedContentMode(WebSettings.MIXED_CONTENT_ALWAYS_ALLOW);
             setting.setAllowContentAccess(false);
             setting.setAllowFileAccessFromFileURLs(true);
@@ -157,10 +167,19 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
             setting.setCacheMode(WebSettings.LOAD_CACHE_ELSE_NETWORK);
             setting.setTextZoom(80);
             String body;
-            if (ResourceUtils.isNightMode(getResources().getConfiguration())) {
-                body = App.HTML_TEMPLATE_DARK.get().replace("@body@", text);
+            String direction;
+            if (getResources().getConfiguration().getLayoutDirection() == View.LAYOUT_DIRECTION_RTL) {
+                direction = "rtl";
             } else {
-                body = App.HTML_TEMPLATE.get().replace("@body@", text);
+                direction = "ltr";
+            }
+            if (text == null) {
+                text = "<center>" + App.getInstance().getString(R.string.list_empty) + "</center>";
+            }
+            if (ResourceUtils.isNightMode(getResources().getConfiguration())) {
+                body = App.HTML_TEMPLATE_DARK.get().replace("@dir@", direction).replace("@body@", text);
+            } else {
+                body = App.HTML_TEMPLATE.get().replace("@dir@", direction).replace("@body@", text);
             }
             view.setWebViewClient(new WebViewClient() {
                 @Override
@@ -209,53 +228,55 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
     }
 
     @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        int id = item.getItemId();
-        if (id == R.id.menu_open_in_browser) {
-            NavUtil.startURL(requireActivity(), "https://modules.lsposed.org/module/" + module.getName());
-        }
-        return super.onOptionsItemSelected(item);
+    public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater) {
+
     }
 
     @Override
-    public void moduleReleasesLoaded(OnlineModule module) {
+    public boolean onMenuItemSelected(@NonNull MenuItem item) {
+        int id = item.getItemId();
+        if (id == R.id.menu_open_in_browser) {
+            NavUtil.startURL(requireActivity(), "https://modules.lsposed.org/module/" + module.getName());
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        RepoLoader.getInstance().removeListener(this);
+        binding = null;
+    }
+
+    @Override
+    public void onModuleReleasesLoaded(OnlineModule module) {
         this.module = module;
+        var repoLoader = RepoLoader.getInstance();
         if (releaseAdapter != null) {
-            runOnUiThread(() -> releaseAdapter.loadItems());
-            if (isResumed() && module.getReleases().size() == 1) {
-                Snackbar.make(binding.snackbar, R.string.module_release_no_more, Snackbar.LENGTH_SHORT).show();
-            }
+            runAsync(releaseAdapter::loadItems);
+        }
+        if ((repoLoader.getReleases(module.getName()) != null ? repoLoader.getReleases(module.getName()).size() : 1) == 1) {
+            showHint(R.string.module_release_no_more, true);
         }
     }
 
     @Override
     public void onThrowable(Throwable t) {
         if (releaseAdapter != null) {
-            runOnUiThread(() -> releaseAdapter.loadItems());
-            if (isResumed()) {
-                Snackbar.make(binding.snackbar, getString(R.string.repo_load_failed, t.getLocalizedMessage()), Snackbar.LENGTH_SHORT).show();
-            }
+            runAsync(releaseAdapter::loadItems);
         }
+        showHint(getString(R.string.repo_load_failed, t.getLocalizedMessage()), true);
     }
 
-    @Override
-    public void onDestroyView() {
-        super.onDestroyView();
-
-        RepoLoader.getInstance().removeListener(this);
-        binding = null;
-    }
-
-    private class InformationAdapter extends RecyclerView.Adapter<InformationAdapter.ViewHolder> {
-        private final OnlineModule module;
+    private class InformationAdapter extends SimpleStatefulAdaptor<InformationAdapter.ViewHolder> {
 
         private int rowCount = 0;
         private int homepageRow = -1;
         private int collaboratorsRow = -1;
         private int sourceUrlRow = -1;
 
-        public InformationAdapter(OnlineModule module) {
-            this.module = module;
+        public InformationAdapter() {
             if (!TextUtils.isEmpty(module.getHomepageUrl())) {
                 homepageRow = rowCount++;
             }
@@ -270,7 +291,7 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
         @NonNull
         @Override
         public InformationAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            return new InformationAdapter.ViewHolder(ItemRepoTitleDescriptionBinding.inflate(getLayoutInflater(), parent, false));
+            return new ViewHolder(ItemRepoTitleDescriptionBinding.inflate(getLayoutInflater(), parent, false));
         }
 
         @Override
@@ -279,13 +300,16 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
                 holder.title.setText(R.string.module_information_homepage);
                 holder.description.setText(module.getHomepageUrl());
             } else if (position == collaboratorsRow) {
-                holder.title.setText(R.string.module_information_collaborators);
                 List<Collaborator> collaborators = module.getCollaborators();
+                if (collaborators == null) return;
+                holder.title.setText(R.string.module_information_collaborators);
                 SpannableStringBuilder sb = new SpannableStringBuilder();
                 ListIterator<Collaborator> iterator = collaborators.listIterator();
                 while (iterator.hasNext()) {
                     Collaborator collaborator = iterator.next();
-                    String name = collaborator.getName() == null ? collaborator.getLogin() : collaborator.getName();
+                    var collaboratorLogin = collaborator.getLogin();
+                    if (collaboratorLogin == null) continue;
+                    String name = collaborator.getName() == null ? collaboratorLogin : collaborator.getName();
                     sb.append(name);
                     CustomTabsURLSpan span = new CustomTabsURLSpan(requireActivity(), String.format("https://github.com/%s", collaborator.getLogin()));
                     sb.setSpan(span, sb.length() - name.length(), sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
@@ -312,7 +336,6 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
                     NavUtil.startURL(requireActivity(), module.getSourceUrl());
                 }
             });
-
         }
 
         @Override
@@ -332,31 +355,90 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
         }
     }
 
-    private class ReleaseAdapter extends RecyclerView.Adapter<ReleaseAdapter.ViewHolder> {
-        private List<Release> items;
-
-        public ReleaseAdapter() {
-            loadItems();
+    public static class DownloadDialog extends DialogFragment {
+        @NonNull
+        @Override
+        public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
+            var args = getArguments();
+            if (args == null) throw new IllegalArgumentException();
+            return new BlurBehindDialogBuilder(requireActivity(), R.style.ThemeOverlay_MaterialAlertDialog_Centered_FullWidthButtons)
+                    .setTitle(R.string.module_release_view_assets)
+                    .setPositiveButton(android.R.string.cancel, null)
+                    .setAdapter(new ArrayAdapter<>(requireActivity(), R.layout.dialog_item, args.getCharSequenceArray("names")),
+                            (dialog, which) -> NavUtil.startURL(requireActivity(), args.getStringArrayList("urls").get(which)))
+                    .create();
         }
 
+        static void create(Activity activity, FragmentManager fm, List<ReleaseAsset> assets) {
+            var f = new DownloadDialog();
+            var bundle = new Bundle();
+
+            var displayNames = new CharSequence[assets.size()];
+            for (int i = 0; i < assets.size(); i++) {
+                var sb = new SpannableStringBuilder(assets.get(i).getName());
+                var count = assets.get(i).getDownloadCount();
+                var countStr = activity.getResources().getQuantityString(R.plurals.module_release_assets_download_count, count, count);
+                var sizeStr = Formatter.formatShortFileSize(activity, assets.get(i).getSize());
+                sb.append('\n').append(sizeStr).append('/').append(countStr);
+                final ForegroundColorSpan foregroundColorSpan = new ForegroundColorSpan(ResourceUtils.resolveColor(activity.getTheme(), android.R.attr.textColorSecondary));
+                final RelativeSizeSpan relativeSizeSpan = new RelativeSizeSpan(0.8f);
+                sb.setSpan(foregroundColorSpan, sb.length() - sizeStr.length() - countStr.length() - 1, sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                sb.setSpan(relativeSizeSpan, sb.length() - sizeStr.length() - countStr.length() - 1, sb.length(), Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                displayNames[i] = sb;
+            }
+            bundle.putCharSequenceArray("names", displayNames);
+            bundle.putStringArrayList("urls", assets.stream().map(ReleaseAsset::getDownloadUrl).collect(Collectors.toCollection(ArrayList::new)));
+            f.setArguments(bundle);
+            f.show(fm, "download");
+        }
+    }
+
+    private class ReleaseAdapter extends EmptyStateRecyclerView.EmptyStateAdapter<ReleaseAdapter.ViewHolder> {
+        private List<Release> items = new ArrayList<>();
+        private final Resources resources = App.getInstance().getResources();
+
+        public ReleaseAdapter() {
+            runAsync(this::loadItems);
+        }
+
+        @SuppressLint("NotifyDataSetChanged")
         public void loadItems() {
-            this.items = module.getReleases();
-            notifyDataSetChanged();
+            var channels = resources.getStringArray(R.array.update_channel_values);
+            var channel = App.getPreferences().getString("update_channel", channels[0]);
+            var releases = RepoLoader.getInstance().getReleases(module.getName());
+            if (releases == null) releases = module.getReleases();
+            List<Release> tmpList;
+            if (channel.equals(channels[0])) {
+                tmpList = releases != null ? releases.parallelStream().filter(t -> {
+                    if (Boolean.TRUE.equals(t.getIsPrerelease())) return false;
+                    var name = t.getName() != null ? t.getName().toLowerCase(LocaleDelegate.getDefaultLocale()) : null;
+                    return !(name != null && name.startsWith("snapshot")) && !(name != null && name.startsWith("nightly"));
+                }).collect(Collectors.toList()) : null;
+            } else if (channel.equals(channels[1])) {
+                tmpList = releases != null ? releases.parallelStream().filter(t -> {
+                    var name = t.getName() != null ? t.getName().toLowerCase(LocaleDelegate.getDefaultLocale()) : null;
+                    return !(name != null && name.startsWith("snapshot")) && !(name != null && name.startsWith("nightly"));
+                }).collect(Collectors.toList()) : null;
+            } else tmpList = releases;
+            runOnUiThread(() -> {
+                items = tmpList;
+                notifyDataSetChanged();
+            });
         }
 
         @NonNull
         @Override
         public ReleaseAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
             if (viewType == 0) {
-                return new ReleaseAdapter.ReleaseViewHolder(ItemRepoReleaseBinding.inflate(getLayoutInflater(), parent, false));
+                return new ReleaseViewHolder(ItemRepoReleaseBinding.inflate(getLayoutInflater(), parent, false));
             } else {
-                return new ReleaseAdapter.LoadmoreViewHolder(ItemRepoLoadmoreBinding.inflate(getLayoutInflater(), parent, false));
+                return new LoadmoreViewHolder(ItemRepoLoadmoreBinding.inflate(getLayoutInflater(), parent, false));
             }
         }
 
         @Override
         public void onBindViewHolder(@NonNull ReleaseAdapter.ViewHolder holder, int position) {
-            if (position == items.size()) {
+            if (holder.getItemViewType() == 1) {
                 holder.progress.setVisibility(View.GONE);
                 holder.title.setVisibility(View.VISIBLE);
                 holder.itemView.setOnClickListener(v -> {
@@ -369,17 +451,15 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
             } else {
                 Release release = items.get(position);
                 holder.title.setText(release.getName());
+                var instant = Instant.parse(release.getPublishedAt());
+                var formatter = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT)
+                        .withLocale(App.getLocale()).withZone(ZoneId.systemDefault());
+                holder.publishedTime.setText(String.format(getString(R.string.module_repo_published_time), formatter.format(instant)));
                 renderGithubMarkdown(holder.description, release.getDescriptionHTML());
                 holder.openInBrowser.setOnClickListener(v -> NavUtil.startURL(requireActivity(), release.getUrl()));
                 List<ReleaseAsset> assets = release.getReleaseAssets();
                 if (assets != null && !assets.isEmpty()) {
-                    holder.viewAssets.setOnClickListener(v -> {
-                        ArrayList<String> names = new ArrayList<>();
-                        assets.forEach(releaseAsset -> names.add(releaseAsset.getName()));
-                        new MaterialAlertDialogBuilder(requireActivity())
-                                .setItems(names.toArray(new String[0]), (dialog, which) -> NavUtil.startURL(requireActivity(), assets.get(which).getDownloadUrl()))
-                                .show();
-                    });
+                    holder.viewAssets.setOnClickListener(v -> DownloadDialog.create(requireActivity(), getParentFragmentManager(), assets));
                 } else {
                     holder.viewAssets.setVisibility(View.GONE);
                 }
@@ -396,8 +476,14 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
             return !module.releasesLoaded && position == getItemCount() - 1 ? 1 : 0;
         }
 
+        @Override
+        public boolean isLoaded() {
+            return module.releasesLoaded;
+        }
+
         class ViewHolder extends RecyclerView.ViewHolder {
             TextView title;
+            TextView publishedTime;
             WebView description;
             MaterialButton openInBrowser;
             MaterialButton viewAssets;
@@ -412,6 +498,7 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
             public ReleaseViewHolder(ItemRepoReleaseBinding binding) {
                 super(binding.getRoot());
                 title = binding.title;
+                publishedTime = binding.publishedTime;
                 description = binding.description;
                 openInBrowser = binding.openInBrowser;
                 viewAssets = binding.viewAssets;
@@ -427,46 +514,27 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
         }
     }
 
-    private class PagerAdapter extends RecyclerView.Adapter<PagerAdapter.ViewHolder> {
+    private static class PagerAdapter extends FragmentStateAdapter {
+
+        public PagerAdapter(@NonNull Fragment fragment) {
+            super(fragment);
+        }
 
         @NonNull
         @Override
-        public PagerAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            if (viewType == 0) {
-                return new PagerAdapter.ReadmeViewHolder(ItemRepoReadmeBinding.inflate(getLayoutInflater(), parent, false));
+        public Fragment createFragment(int position) {
+            Bundle bundle = new Bundle();
+            bundle.putInt("position", position);
+            Fragment f;
+            if (position == 0) {
+                f = new ReadmeFragment();
+            } else if (position == 1) {
+                f = new RecyclerviewFragment();
             } else {
-                return new PagerAdapter.RecyclerviewBinding(ItemRepoRecyclerviewBinding.inflate(getLayoutInflater(), parent, false));
+                f = new RecyclerviewFragment();
             }
-        }
-
-        @Override
-        public void onBindViewHolder(@NonNull PagerAdapter.ViewHolder holder, int position) {
-            switch (position) {
-                case 0:
-                    holder.scrollView.getBorderViewDelegate().setBorderVisibilityChangedListener((top, oldTop, bottom, oldBottom) -> binding.appBar.setLifted(!top));
-                    holder.scrollView.setTag(position);
-                    if (module != null)
-                        renderGithubMarkdown(holder.webView, module.getReadmeHTML());
-                    break;
-                case 1:
-                case 2:
-                    if (position == 1) {
-                        holder.recyclerView.setAdapter(releaseAdapter = new ReleaseAdapter());
-                    } else {
-                        holder.recyclerView.setAdapter(new InformationAdapter(module));
-                    }
-                    int height = ResourceUtils.resolveDimensionPixelOffset(requireActivity().getTheme(), androidx.appcompat.R.attr.actionBarSize, 0)
-                            + getResources().getDimensionPixelOffset(R.dimen.tab_layout_height);
-                    WindowInsetsHelperKt.setInitialPadding(holder.recyclerView, 0, height, 0, 0);
-                    holder.recyclerView.setTag(position);
-                    holder.recyclerView.setLayoutManager(new LinearLayoutManager(requireActivity()));
-                    holder.recyclerView.getBorderViewDelegate().setBorderVisibilityChangedListener((top, oldTop, bottom, oldBottom) -> binding.appBar.setLifted(!top));
-                    var insets = requireActivity().getWindow().getDecorView().getRootWindowInsets();
-                    if (insets != null)
-                        holder.recyclerView.onApplyWindowInsets(insets);
-                    RecyclerViewKt.fixEdgeEffect(holder.recyclerView, false, true);
-                    break;
-            }
+            f.setArguments(bundle);
+            return f;
         }
 
         @Override
@@ -479,29 +547,116 @@ public class RepoItemFragment extends BaseFragment implements RepoLoader.Listene
             return position == 0 ? 0 : 1;
         }
 
-        class ViewHolder extends RecyclerView.ViewHolder {
-            WebView webView;
-            BorderNestedScrollView scrollView;
-            BorderRecyclerView recyclerView;
+        @Override
+        public long getItemId(int position) {
+            return position;
+        }
+    }
 
-            public ViewHolder(@NonNull View itemView) {
-                super(itemView);
+    public static abstract class BorderFragment extends BaseFragment {
+        BorderView borderView;
+
+        void attachListeners() {
+            var parent = getParentFragment();
+            if (parent instanceof RepoItemFragment) {
+                var repoItemFragment = (RepoItemFragment) parent;
+                borderView.getBorderViewDelegate().setBorderVisibilityChangedListener((top, oldTop, bottom, oldBottom) -> repoItemFragment.binding.appBar.setLifted(!top));
+                repoItemFragment.binding.appBar.setLifted(!borderView.getBorderViewDelegate().isShowingTopBorder());
+                repoItemFragment.binding.toolbar.setOnClickListener(v -> {
+                    repoItemFragment.binding.appBar.setExpanded(true, true);
+                    scrollToTop();
+                });
             }
         }
 
-        class ReadmeViewHolder extends PagerAdapter.ViewHolder {
-            public ReadmeViewHolder(ItemRepoReadmeBinding binding) {
-                super(binding.getRoot());
-                webView = binding.readme;
-                scrollView = binding.scrollView;
-            }
+        abstract void scrollToTop();
+
+        void detachListeners() {
+            borderView.getBorderViewDelegate().setBorderVisibilityChangedListener(null);
         }
 
-        class RecyclerviewBinding extends PagerAdapter.ViewHolder {
-            public RecyclerviewBinding(ItemRepoRecyclerviewBinding binding) {
-                super(binding.getRoot());
-                recyclerView = binding.recyclerView;
+        @Override
+        public void onResume() {
+            super.onResume();
+            attachListeners();
+        }
+
+        @Override
+        public void onStart() {
+            super.onStart();
+            attachListeners();
+        }
+
+        @Override
+        public void onStop() {
+            super.onStop();
+            detachListeners();
+        }
+
+        @Override
+        public void onPause() {
+            super.onPause();
+            detachListeners();
+        }
+    }
+
+    public static class ReadmeFragment extends BorderFragment {
+        ItemRepoReadmeBinding binding;
+
+        @Nullable
+        @Override
+        public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+            var parent = getParentFragment();
+            if (!(parent instanceof RepoItemFragment)) {
+                if (!safeNavigate(R.id.action_repo_item_fragment_to_repo_fragment)) {
+                    safeNavigate(R.id.repo_nav);
+                }
+                return null;
             }
+            var repoItemFragment = (RepoItemFragment) parent;
+            binding = ItemRepoReadmeBinding.inflate(getLayoutInflater(), container, false);
+            repoItemFragment.renderGithubMarkdown(binding.readme, repoItemFragment.module.getReadmeHTML());
+            borderView = binding.scrollView;
+            return binding.getRoot();
+        }
+
+        @Override
+        void scrollToTop() {
+            binding.scrollView.fullScroll(ScrollView.FOCUS_UP);
+        }
+    }
+
+    public static class RecyclerviewFragment extends BorderFragment {
+        ItemRepoRecyclerviewBinding binding;
+        RecyclerView.Adapter<?> adapter;
+
+        @Override
+        void scrollToTop() {
+            binding.recyclerView.smoothScrollToPosition(0);
+        }
+
+        public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+            var arguments = getArguments();
+            var parent = getParentFragment();
+            if (arguments == null || !(parent instanceof RepoItemFragment)) {
+                if (!safeNavigate(R.id.action_repo_item_fragment_to_repo_fragment)) {
+                    safeNavigate(R.id.repo_nav);
+                }
+                return null;
+            }
+            var repoItemFragment = (RepoItemFragment) parent;
+            var position = arguments.getInt("position", 0);
+            if (position == 1)
+                adapter = repoItemFragment.releaseAdapter;
+            else if (position == 2)
+                adapter = repoItemFragment.informationAdapter;
+            else return null;
+            binding = ItemRepoRecyclerviewBinding.inflate(getLayoutInflater(), container, false);
+            binding.recyclerView.setAdapter(adapter);
+            binding.recyclerView.setLayoutManager(new LinearLayoutManager(requireActivity()));
+            RecyclerViewKt.fixEdgeEffect(binding.recyclerView, false, true);
+            borderView = binding.recyclerView;
+            return binding.getRoot();
         }
     }
 }
