@@ -19,9 +19,8 @@
 
 package org.lsposed.manager.ui.fragment;
 
+import android.content.ActivityNotFoundException;
 import android.content.Context;
-import android.content.res.Configuration;
-import android.content.res.Resources.NotFoundException;
 import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
@@ -29,29 +28,31 @@ import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatDelegate;
 import androidx.core.text.HtmlCompat;
 import androidx.preference.Preference;
-import androidx.preference.SwitchPreference;
+import androidx.preference.PreferenceFragmentCompat;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.google.android.material.color.DynamicColors;
-import com.google.android.material.snackbar.Snackbar;
-import com.takisoft.preferencex.PreferenceFragmentCompat;
 
 import org.lsposed.manager.App;
 import org.lsposed.manager.BuildConfig;
 import org.lsposed.manager.ConfigManager;
 import org.lsposed.manager.R;
 import org.lsposed.manager.databinding.FragmentSettingsBinding;
+import org.lsposed.manager.repo.RepoLoader;
 import org.lsposed.manager.ui.activity.MainActivity;
 import org.lsposed.manager.util.BackupUtils;
+import org.lsposed.manager.util.CloudflareDNS;
+import org.lsposed.manager.util.LangList;
 import org.lsposed.manager.util.NavUtil;
+import org.lsposed.manager.util.ShortcutUtil;
 import org.lsposed.manager.util.ThemeUtil;
 
 import java.time.LocalDateTime;
@@ -59,8 +60,8 @@ import java.util.ArrayList;
 import java.util.Locale;
 
 import rikka.core.util.ResourceUtils;
-import rikka.material.app.DayNightDelegate;
 import rikka.material.app.LocaleDelegate;
+import rikka.material.preference.MaterialSwitchPreference;
 import rikka.preference.SimpleMenuPreference;
 import rikka.recyclerview.RecyclerViewKt;
 import rikka.widget.borderview.BorderRecyclerView;
@@ -72,12 +73,16 @@ public class SettingsFragment extends BaseFragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         binding = FragmentSettingsBinding.inflate(inflater, container, false);
-        binding.getRoot().bringChildToFront(binding.appBar);
         binding.appBar.setLiftable(true);
-        setupToolbar(binding.toolbar, R.string.Settings);
+        setupToolbar(binding.toolbar, binding.clickView, R.string.Settings);
+        binding.toolbar.setNavigationIcon(null);
         if (savedInstanceState == null) {
-            getChildFragmentManager().beginTransaction()
-                    .add(R.id.container, new PreferenceFragment()).commit();
+            getChildFragmentManager().beginTransaction().add(R.id.setting_container, new PreferenceFragment()).commitNow();
+        }
+        if (ConfigManager.isBinderAlive()) {
+            binding.toolbar.setSubtitle(String.format(LocaleDelegate.getDefaultLocale(), "%s (%d) - %s", ConfigManager.getXposedVersionName(), ConfigManager.getXposedVersionCode(), ConfigManager.getApi()));
+        } else {
+            binding.toolbar.setSubtitle(String.format(LocaleDelegate.getDefaultLocale(), "%s (%d) - %s", BuildConfig.VERSION_NAME, BuildConfig.VERSION_CODE, getString(R.string.not_installed)));
         }
         return binding.getRoot();
     }
@@ -92,38 +97,28 @@ public class SettingsFragment extends BaseFragment {
     public static class PreferenceFragment extends PreferenceFragmentCompat {
         private SettingsFragment parentFragment;
 
-        ActivityResultLauncher<String> backupLauncher = registerForActivityResult(new ActivityResultContracts.CreateDocument(),
-                uri -> {
-                    if (uri == null || parentFragment == null) return;
-                    parentFragment.runAsync(() -> {
-                        try {
-                            BackupUtils.backup(uri);
-                        } catch (Exception e) {
-                            var text = App.getInstance().getString(R.string.settings_backup_failed2, e.getMessage());
-                            if (parentFragment != null && parentFragment.binding != null && isResumed()) {
-                                Snackbar.make(parentFragment.binding.snackbar, text, Snackbar.LENGTH_LONG).show();
-                            } else {
-                                Toast.makeText(App.getInstance(), text, Toast.LENGTH_LONG).show();
-                            }
-                        }
-                    });
-                });
-        ActivityResultLauncher<String[]> restoreLauncher = registerForActivityResult(new ActivityResultContracts.OpenDocument(),
-                uri -> {
-                    if (uri == null || parentFragment == null) return;
-                    parentFragment.runAsync(() -> {
-                        try {
-                            BackupUtils.restore(uri);
-                        } catch (Exception e) {
-                            var text = App.getInstance().getString(R.string.settings_restore_failed2, e.getMessage());
-                            if (parentFragment != null && parentFragment.binding != null && isResumed()) {
-                                Snackbar.make(parentFragment.binding.snackbar, text, Snackbar.LENGTH_LONG).show();
-                            } else {
-                                Toast.makeText(App.getInstance(), text, Toast.LENGTH_LONG).show();
-                            }
-                        }
-                    });
-                });
+        ActivityResultLauncher<String> backupLauncher = registerForActivityResult(new ActivityResultContracts.CreateDocument("application/gzip"), uri -> {
+            if (uri == null || parentFragment == null) return;
+            parentFragment.runAsync(() -> {
+                try {
+                    BackupUtils.backup(uri);
+                } catch (Exception e) {
+                    var text = App.getInstance().getString(R.string.settings_backup_failed2, e.getMessage());
+                    parentFragment.showHint(text, false);
+                }
+            });
+        });
+        ActivityResultLauncher<String[]> restoreLauncher = registerForActivityResult(new ActivityResultContracts.OpenDocument(), uri -> {
+            if (uri == null || parentFragment == null) return;
+            parentFragment.runAsync(() -> {
+                try {
+                    BackupUtils.restore(uri);
+                } catch (Exception e) {
+                    var text = App.getInstance().getString(R.string.settings_restore_failed2, e.getMessage());
+                    parentFragment.showHint(text, false);
+                }
+            });
+        });
 
         @Override
         public void onAttach(@NonNull Context context) {
@@ -139,34 +134,74 @@ public class SettingsFragment extends BaseFragment {
             parentFragment = null;
         }
 
+        private boolean setNotificationPreferenceEnabled(MaterialSwitchPreference notificationPreference, boolean preferenceEnabled) {
+            var notificationEnabled = ConfigManager.enableStatusNotification();
+            if (notificationPreference != null) {
+                notificationPreference.setEnabled(!notificationEnabled || preferenceEnabled);
+                notificationPreference.setSummaryOn(preferenceEnabled ?
+                        notificationPreference.getContext().getString(R.string.settings_enable_status_notification_summary) :
+                        notificationPreference.getContext().getString(R.string.settings_enable_status_notification_summary) + "\n" +
+                                notificationPreference.getContext().getString(R.string.disable_status_notification_error));
+            }
+            return notificationEnabled;
+        }
+
         @Override
-        public void onCreatePreferencesFix(Bundle savedInstanceState, String rootKey) {
+        public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
             final String SYSTEM = "SYSTEM";
 
             addPreferencesFromResource(R.xml.prefs);
 
             boolean installed = ConfigManager.isBinderAlive();
-            SwitchPreference prefVerboseLogs = findPreference("disable_verbose_log");
+            MaterialSwitchPreference prefVerboseLogs = findPreference("disable_verbose_log");
             if (prefVerboseLogs != null) {
                 prefVerboseLogs.setEnabled(!BuildConfig.DEBUG && installed);
                 prefVerboseLogs.setChecked(!installed || !ConfigManager.isVerboseLogEnabled());
-                prefVerboseLogs.setOnPreferenceChangeListener((preference, newValue) ->
-                        ConfigManager.setVerboseLogEnabled(!(boolean) newValue));
+                prefVerboseLogs.setOnPreferenceChangeListener((preference, newValue) -> ConfigManager.setVerboseLogEnabled(!(boolean) newValue));
             }
 
-            SwitchPreference prefEnableResources = findPreference("enable_resources");
-            if (prefEnableResources != null) {
-                prefEnableResources.setEnabled(installed);
-                prefEnableResources.setChecked(installed && ConfigManager.isResourceHookEnabled());
-                prefEnableResources.setOnPreferenceChangeListener((preference, newValue) -> ConfigManager.setResourceHookEnabled((boolean) newValue));
+            MaterialSwitchPreference prefDexObfuscate = findPreference("enable_dex_obfuscate");
+            if (prefDexObfuscate != null) {
+                prefDexObfuscate.setEnabled(installed);
+                prefDexObfuscate.setChecked(!installed || ConfigManager.isDexObfuscateEnabled());
+                prefDexObfuscate.setOnPreferenceChangeListener((preference, newValue) -> {
+                    parentFragment.showHint(R.string.reboot_required, true, R.string.reboot, v -> ConfigManager.reboot());
+                    return ConfigManager.setDexObfuscateEnabled((boolean) newValue);
+                });
             }
 
-            SwitchPreference prefEnableShortcut = findPreference("enable_auto_add_shortcut");
-            if (prefEnableShortcut != null) {
-                prefEnableShortcut.setEnabled(installed);
-                prefEnableShortcut.setVisible(!App.isParasitic());
-                prefEnableShortcut.setChecked(installed && ConfigManager.isAddShortcut());
-                prefEnableShortcut.setOnPreferenceChangeListener((preference, newValue) -> ConfigManager.setAddShortcut((boolean) newValue));
+            MaterialSwitchPreference notificationPreference = findPreference("enable_status_notification");
+            if (notificationPreference != null) {
+                notificationPreference.setVisible(installed);
+                if (installed) {
+                    notificationPreference.setChecked(setNotificationPreferenceEnabled(notificationPreference, !App.isParasitic || ShortcutUtil.isLaunchShortcutPinned()));
+                }
+                notificationPreference.setOnPreferenceChangeListener((p, v) -> {
+                    var succeeded = ConfigManager.setEnableStatusNotification((boolean) v);
+                    if ((boolean) v && App.isParasitic && !ShortcutUtil.isLaunchShortcutPinned()) {
+                        setNotificationPreferenceEnabled(notificationPreference, false);
+                    }
+                    return succeeded;
+                });
+            }
+
+            Preference shortcut = findPreference("add_shortcut");
+            if (shortcut != null) {
+                shortcut.setVisible(App.isParasitic);
+                if (!ShortcutUtil.isRequestPinShortcutSupported(requireContext())) {
+                    shortcut.setEnabled(false);
+                    shortcut.setSummary(R.string.settings_unsupported_pin_shortcut_summary);
+                }
+                shortcut.setOnPreferenceClickListener(preference -> {
+                    if (!ShortcutUtil.requestPinLaunchShortcut(() -> {
+                        setNotificationPreferenceEnabled(notificationPreference, true);
+                        App.getPreferences().edit().putBoolean("never_show_welcome", true).apply();
+                        parentFragment.showHint(R.string.settings_shortcut_pinned_hint, false);
+                    })) {
+                        parentFragment.showHint(R.string.settings_unsupported_pin_shortcut_summary, true);
+                    }
+                    return true;
+                });
             }
 
             Preference backup = findPreference("backup");
@@ -174,9 +209,13 @@ public class SettingsFragment extends BaseFragment {
                 backup.setEnabled(installed);
                 backup.setOnPreferenceClickListener(preference -> {
                     LocalDateTime now = LocalDateTime.now();
-                    backupLauncher.launch(String.format(Locale.ROOT,
-                            "LSPosed_%s.lsp", now.toString()));
-                    return true;
+                    try {
+                        backupLauncher.launch(String.format(LocaleDelegate.getDefaultLocale(), "LSPosed_%s.lsp", now.toString()));
+                        return true;
+                    } catch (ActivityNotFoundException e) {
+                        parentFragment.showHint(R.string.enable_documentui, true);
+                        return false;
+                    }
                 });
             }
 
@@ -184,8 +223,13 @@ public class SettingsFragment extends BaseFragment {
             if (restore != null) {
                 restore.setEnabled(installed);
                 restore.setOnPreferenceClickListener(preference -> {
-                    restoreLauncher.launch(new String[]{"*/*"});
-                    return true;
+                    try {
+                        restoreLauncher.launch(new String[]{"*/*"});
+                        return true;
+                    } catch (ActivityNotFoundException e) {
+                        parentFragment.showHint(R.string.enable_documentui, true);
+                        return false;
+                    }
                 });
             }
 
@@ -193,11 +237,7 @@ public class SettingsFragment extends BaseFragment {
             if (theme != null) {
                 theme.setOnPreferenceChangeListener((preference, newValue) -> {
                     if (!App.getPreferences().getString("dark_theme", ThemeUtil.MODE_NIGHT_FOLLOW_SYSTEM).equals(newValue)) {
-                        DayNightDelegate.setDefaultNightMode(ThemeUtil.getDarkTheme((String) newValue));
-                        MainActivity activity = (MainActivity) getActivity();
-                        if (activity != null) {
-                            activity.restart();
-                        }
+                        AppCompatDelegate.setDefaultNightMode(ThemeUtil.getDarkTheme((String) newValue));
                     }
                     return true;
                 });
@@ -225,18 +265,16 @@ public class SettingsFragment extends BaseFragment {
                 });
             }
 
-            SwitchPreference prefShowHiddenIcons = findPreference("show_hidden_icon_apps_enabled");
+            MaterialSwitchPreference prefShowHiddenIcons = findPreference("show_hidden_icon_apps_enabled");
             if (prefShowHiddenIcons != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
                 if (ConfigManager.isBinderAlive()) {
                     prefShowHiddenIcons.setEnabled(true);
-                    prefShowHiddenIcons.setOnPreferenceChangeListener((preference, newValue) ->
-                            ConfigManager.setHiddenIcon(!(boolean) newValue));
+                    prefShowHiddenIcons.setOnPreferenceChangeListener((preference, newValue) -> ConfigManager.setHiddenIcon(!(boolean) newValue));
                 }
-                prefShowHiddenIcons.setChecked(Settings.Global.getInt(
-                        requireActivity().getContentResolver(), "show_hidden_icon_apps_enabled", 1) != 0);
+                prefShowHiddenIcons.setChecked(Settings.Global.getInt(requireActivity().getContentResolver(), "show_hidden_icon_apps_enabled", 1) != 0);
             }
 
-            SwitchPreference prefFollowSystemAccent = findPreference("follow_system_accent");
+            MaterialSwitchPreference prefFollowSystemAccent = findPreference("follow_system_accent");
             if (prefFollowSystemAccent != null && DynamicColors.isDynamicColorAvailable()) {
                 if (primary_color != null) {
                     primary_color.setVisible(!prefFollowSystemAccent.isChecked());
@@ -251,38 +289,59 @@ public class SettingsFragment extends BaseFragment {
                 });
             }
 
+            MaterialSwitchPreference prefDoH = findPreference("doh");
+            if (prefDoH != null) {
+                var dns = (CloudflareDNS) App.getOkHttpClient().dns();
+                if (!dns.noProxy) {
+                    prefDoH.setEnabled(false);
+                    prefDoH.setVisible(false);
+                    var group = prefDoH.getParent();
+                    assert group != null;
+                    group.setVisible(false);
+                }
+                prefDoH.setOnPreferenceChangeListener((p, v) -> {
+                    dns.DoH = (boolean) v;
+                    return true;
+                });
+            }
+
             SimpleMenuPreference language = findPreference("language");
             if (language != null) {
+                var tag = language.getValue();
+                var userLocale = App.getLocale();
+                var entries = new ArrayList<CharSequence>();
+                var lstLang = LangList.LOCALES;
+                for (var lang : lstLang) {
+                    if (lang.equals(SYSTEM)) {
+                        entries.add(getString(rikka.core.R.string.follow_system));
+                        continue;
+                    }
+                    var locale = Locale.forLanguageTag(lang);
+                    entries.add(HtmlCompat.fromHtml(locale.getDisplayName(locale), HtmlCompat.FROM_HTML_MODE_LEGACY));
+                }
+                language.setEntries(entries.toArray(new CharSequence[0]));
+                language.setEntryValues(lstLang);
+                if (TextUtils.isEmpty(tag) || SYSTEM.equals(tag)) {
+                    language.setSummary(getString(rikka.core.R.string.follow_system));
+                } else {
+                    var locale = Locale.forLanguageTag(tag);
+                    language.setSummary(!TextUtils.isEmpty(locale.getScript()) ? locale.getDisplayScript(userLocale) : locale.getDisplayName(userLocale));
+                }
                 language.setOnPreferenceChangeListener((preference, newValue) -> {
-                    var locale = SYSTEM.equals(newValue) ? LocaleDelegate.getSystemLocale() : Locale.forLanguageTag((String) newValue);
+                    var app = App.getInstance();
+                    var locale = App.getLocale((String) newValue);
+                    var res = app.getResources();
+                    var config = res.getConfiguration();
+                    config.setLocale(locale);
                     LocaleDelegate.setDefaultLocale(locale);
+                    //noinspection deprecation
+                    res.updateConfiguration(config, res.getDisplayMetrics());
                     MainActivity activity = (MainActivity) getActivity();
                     if (activity != null) {
                         activity.restart();
                     }
                     return true;
                 });
-                var tag = language.getValue();
-                var userLocale = App.getLocale();
-                var entries = new ArrayList<CharSequence>();
-                entries.add(language.getEntries()[0]);
-                var lstLang = getAppLanguages(requireContext(), R.string.Settings);
-                for (var lang : lstLang) {
-                    var locale = Locale.forLanguageTag(lang);
-                    entries.add(HtmlCompat.fromHtml(String.format("%s - %s",
-                            !TextUtils.isEmpty(locale.getScript()) ? locale.getDisplayScript(locale) : locale.getDisplayName(locale),
-                            !TextUtils.isEmpty(locale.getScript()) ? locale.getDisplayScript(userLocale) : locale.getDisplayName(userLocale)
-                    ), HtmlCompat.FROM_HTML_MODE_LEGACY));
-                }
-                language.setEntries(entries.toArray(new CharSequence[0]));
-                lstLang.add(0, SYSTEM);
-                language.setEntryValues(lstLang.toArray(new CharSequence[0]));
-                if (TextUtils.isEmpty(tag) || SYSTEM.equals(tag)) {
-                    language.setSummary(getString(rikka.material.R.string.follow_system));
-                } else {
-                    var locale = Locale.forLanguageTag(tag);
-                    language.setSummary(!TextUtils.isEmpty(locale.getScript()) ? locale.getDisplayScript(userLocale) : locale.getDisplayName(userLocale));
-                }
             }
 
             Preference translation = findPreference("translation");
@@ -303,48 +362,32 @@ public class SettingsFragment extends BaseFragment {
                     translation_contributors.setSummary(translators);
                 }
             }
+            SimpleMenuPreference channel = findPreference("update_channel");
+            if (channel != null) {
+                channel.setOnPreferenceChangeListener((preference, newValue) -> {
+                    var repoLoader = RepoLoader.getInstance();
+                    repoLoader.updateLatestVersion(String.valueOf(newValue));
+                    return true;
+                });
+            }
         }
 
+        @NonNull
         @Override
-        public RecyclerView onCreateRecyclerView(LayoutInflater inflater, ViewGroup parent, Bundle savedInstanceState) {
+        public RecyclerView onCreateRecyclerView(@NonNull LayoutInflater inflater, @NonNull ViewGroup parent, Bundle savedInstanceState) {
             BorderRecyclerView recyclerView = (BorderRecyclerView) super.onCreateRecyclerView(inflater, parent, savedInstanceState);
             RecyclerViewKt.fixEdgeEffect(recyclerView, false, true);
-            recyclerView.getBorderViewDelegate().setBorderVisibilityChangedListener((top, oldTop, bottom, oldBottom) -> {
-                SettingsFragment fragment = (SettingsFragment) getParentFragment();
-                if (fragment != null) {
-                    fragment.binding.appBar.setLifted(!top);
-                }
-            });
+            recyclerView.getBorderViewDelegate().setBorderVisibilityChangedListener((top, oldTop, bottom, oldBottom) -> parentFragment.binding.appBar.setLifted(!top));
+            var fragment = getParentFragment();
+            if (fragment instanceof SettingsFragment settingsFragment) {
+                View.OnClickListener l = v -> {
+                    settingsFragment.binding.appBar.setExpanded(true, true);
+                    recyclerView.smoothScrollToPosition(0);
+                };
+                settingsFragment.binding.toolbar.setOnClickListener(l);
+                settingsFragment.binding.clickView.setOnClickListener(l);
+            }
             return recyclerView;
-        }
-
-        private ArrayList<String> getAppLanguages(Context ctx, int id) {
-            Configuration conf = ctx.getResources().getConfiguration();
-            Locale originalLocale = conf.getLocales().get(0);
-            conf.setLocale(Locale.ENGLISH);
-
-            var lstLang = new ArrayList<String>();
-            lstLang.add(Locale.ENGLISH.getLanguage());
-
-            final String reference;
-            try {
-                reference = ctx.createConfigurationContext(conf).getString(id);
-            } catch(NotFoundException nfe) {
-                return lstLang; // return only english
-            }
-
-            for (String loc : ctx.getAssets().getLocales()) {
-                if (loc.isEmpty()) {
-                    continue;
-                }
-                Locale locale = Locale.forLanguageTag(loc);
-                conf.setLocale(locale);
-                if (!lstLang.contains(loc) && !reference.equals(ctx.createConfigurationContext(conf).getString(id))) {
-                    lstLang.add(loc);
-                }
-            }
-            conf.setLocale(originalLocale);
-            return lstLang;
         }
     }
 }
