@@ -1,3 +1,22 @@
+/*
+ * This file is part of LSPosed.
+ *
+ * LSPosed is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LSPosed is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LSPosed.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Copyright (C) 2022 LSPosed Contributors
+ */
+
 package org.lsposed.manager.ui.widget;
 
 import android.annotation.SuppressLint;
@@ -10,8 +29,9 @@ import android.webkit.WebView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.core.view.ScrollingView;
-import androidx.viewpager2.widget.ViewPager2;
+import androidx.recyclerview.widget.RecyclerView;
+
+import rikka.widget.borderview.BorderRecyclerView;
 
 public class ScrollWebView extends WebView {
     public ScrollWebView(@NonNull Context context) {
@@ -42,7 +62,7 @@ public class ScrollWebView extends WebView {
 
     @Override
     protected void onOverScrolled(int scrollX, int scrollY, boolean clampedX, boolean clampedY) {
-        if (clampedX || clampedY) {
+        if (clampedX) {
             var viewParent = findViewParentIfNeeds(this);
             if (viewParent != null) viewParent.requestDisallowInterceptTouchEvent(false);
         }
@@ -52,7 +72,7 @@ public class ScrollWebView extends WebView {
     private static ViewParent findViewParentIfNeeds(View v) {
         var parent = v.getParent();
         if (parent == null) return null;
-        if (parent instanceof ViewPager2 || parent instanceof ScrollingView) {
+        if (parent instanceof RecyclerView && !(parent instanceof BorderRecyclerView)) {
             return parent;
         } else if (parent instanceof View) {
             return findViewParentIfNeeds((View) parent);
