@@ -23,6 +23,7 @@ package android.content.res;
 import static org.lsposed.lspd.nativebridge.ResourcesHook.rewriteXmlReferencesNative;
 import static de.robv.android.xposed.XposedHelpers.decrementMethodDepth;
 import static de.robv.android.xposed.XposedHelpers.findAndHookMethod;
+import static de.robv.android.xposed.XposedHelpers.getBooleanField;
 import static de.robv.android.xposed.XposedHelpers.getLongField;
 import static de.robv.android.xposed.XposedHelpers.getObjectField;
 import static de.robv.android.xposed.XposedHelpers.incrementMethodDepth;
@@ -32,8 +33,10 @@ import android.content.pm.PackageParser;
 import android.content.pm.PackageParser.PackageParserException;
 import android.graphics.Color;
 import android.graphics.Movie;
+import android.graphics.Typeface;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
+import android.os.Build;
 import android.os.StrictMode;
 import android.text.Html;
 import android.util.AttributeSet;
@@ -43,6 +46,9 @@ import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 
 import org.xmlpull.v1.XmlPullParser;
 
@@ -57,6 +63,7 @@ import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XC_MethodHook.MethodHookParam;
 import de.robv.android.xposed.XposedBridge;
 import de.robv.android.xposed.XposedBridge.CopyOnWriteSortedSet;
+import de.robv.android.xposed.XposedInit;
 import de.robv.android.xposed.callbacks.XC_LayoutInflated;
 import de.robv.android.xposed.callbacks.XC_LayoutInflated.LayoutInflatedParam;
 import de.robv.android.xposed.callbacks.XCallback;
@@ -75,6 +82,10 @@ public class XResources extends XResourcesSuperClass {
 	private static final SparseArray<HashMap<String, Object>> sReplacements = new SparseArray<>();
 	private static final SparseArray<HashMap<String, ResourceNames>> sResourceNames = new SparseArray<>();
 
+	// A resource ID is a 32 bit number of the form: PPTTNNNN. PP is the package the resource is for;
+	// TT is the type of the resource;
+	// NNNN is the name of the resource in that type.
+	// For applications resources, PP is always 0x7f.
 	private static final byte[] sSystemReplacementsCache = new byte[256]; // bitmask: 0x000700ff => 2048 bit => 256 bytes
 	private byte[] mReplacementsCache; // bitmask: 0x0007007f => 1024 bit => 128 bytes
 	private static final HashMap<String, byte[]> sReplacementsCacheMap = new HashMap<>();
@@ -84,34 +95,17 @@ public class XResources extends XResourcesSuperClass {
 	private static final WeakHashMap<XmlResourceParser, XMLInstanceDetails> sXmlInstanceDetails = new WeakHashMap<>();
 
 	private static final String EXTRA_XML_INSTANCE_DETAILS = "xmlInstanceDetails";
-	private static final ThreadLocal<LinkedList<MethodHookParam>> sIncludedLayouts = new ThreadLocal<LinkedList<MethodHookParam>>() {
-		@Override
-		protected LinkedList<MethodHookParam> initialValue() {
-			return new LinkedList<>();
-		}
-	};
+	private static final ThreadLocal<LinkedList<MethodHookParam>> sIncludedLayouts = ThreadLocal.withInitial(() -> new LinkedList<>());
 
 	private static final HashMap<String, Long> sResDirLastModified = new HashMap<>();
 	private static final HashMap<String, String> sResDirPackageNames = new HashMap<>();
 	private static ThreadLocal<Object> sLatestResKey = null;
 
-	private boolean mIsObjectInited;
 	private String mResDir;
 	private String mPackageName;
 
-	public XResources(ClassLoader classLoader) {
+	public XResources(ClassLoader classLoader, String resDir) {
 		super(classLoader);
-	}
-
-	/** Dummy, will never be called (objects are transferred to this class only). */
-//	private XResources() {
-//		throw new UnsupportedOperationException();
-//	}
-
-	/** @hide */
-	public void initObject(String resDir) {
-		if (mIsObjectInited)
-			throw new IllegalStateException("Object has already been initialized");
 
 		this.mResDir = resDir;
 		this.mPackageName = getPackageName(resDir);
@@ -121,9 +115,12 @@ public class XResources extends XResourcesSuperClass {
 				mReplacementsCache = sReplacementsCacheMap.computeIfAbsent(resDir, k -> new byte[128]);
 			}
 		}
-
-		this.mIsObjectInited = true;
 	}
+
+	/** Dummy, will never be called (objects are transferred to this class only). */
+//	private XResources() {
+//		throw new UnsupportedOperationException();
+//	}
 
 	/** @hide */
 	public boolean isFirstLoad() {
@@ -162,6 +159,7 @@ public class XResources extends XResourcesSuperClass {
 	/**
 	 * Returns the name of the package that these resources belong to, or "android" for system resources.
 	 */
+	@NonNull
 	public String getPackageName() {
 		return mPackageName;
 	}
@@ -567,6 +565,13 @@ public class XResources extends XResourcesSuperClass {
 
 	private static void setReplacement(int id, Object replacement, XResources res) {
 		String resDir = (res != null) ? res.mResDir : null;
+		if (res == null) {
+			try {
+				XposedInit.hookResources();
+			} catch (Throwable throwable) {
+				throw new IllegalStateException("Failed to initialize resources hook");
+			}
+		}
 		if (id == 0)
 			throw new IllegalArgumentException("id 0 is not an allowed resource identifier");
 		else if (resDir == null && id >= 0x7f000000)
@@ -853,6 +858,35 @@ public class XResources extends XResourcesSuperClass {
 	}
 
 	/** @hide */
+	@RequiresApi(Build.VERSION_CODES.Q)
+	@Override
+	public float getFloat(int id) {
+		Object replacement = getReplacement(id);
+		if (replacement instanceof Float) {
+			return (Float) replacement;
+		} else if (replacement instanceof XResForwarder) {
+			Resources repRes = ((XResForwarder) replacement).getResources();
+			int repId = ((XResForwarder) replacement).getId();
+			return repRes.getFloat(repId);
+		}
+		return super.getFloat(id);
+	}
+
+	/** @hide */
+	@Override
+	public Typeface getFont(int id) {
+		Object replacement = getReplacement(id);
+		if (replacement instanceof Typeface) {
+			return (Typeface) replacement;
+		} else if (replacement instanceof XResForwarder) {
+			Resources repRes = ((XResForwarder) replacement).getResources();
+			int repId = ((XResForwarder) replacement).getId();
+			return repRes.getFont(repId);
+		}
+		return super.getFont(id);
+	}
+
+	/** @hide */
 	@Override
 	public float getFraction(int id, int base, int pbase) {
 		Object replacement = getReplacement(id);
@@ -1045,6 +1079,38 @@ public class XResources extends XResourcesSuperClass {
 			return repRes.getTextArray(repId);
 		}
 		return super.getTextArray(id);
+	}
+
+	/** @hide */
+	@Override
+	public void getValue(int id, TypedValue outValue, boolean resolveRefs) throws NotFoundException {
+		Object replacement = getReplacement(id);
+		if (replacement instanceof XResForwarder) {
+			Resources repRes = ((XResForwarder) replacement).getResources();
+			int repId = ((XResForwarder) replacement).getId();
+			repRes.getValue(repId, outValue, resolveRefs);
+		} else {
+			if (replacement != null) {
+				XposedBridge.log("Replacement of resource ID #0x" + Integer.toHexString(id) + " escaped because of deprecated replacement. Please use XResForwarder instead.");
+			}
+			super.getValue(id, outValue, resolveRefs);
+		}
+	}
+
+	/** @hide */
+	@Override
+	public void getValueForDensity(int id, int density, TypedValue outValue, boolean resolveRefs) throws NotFoundException {
+		Object replacement = getReplacement(id);
+		if (replacement instanceof XResForwarder) {
+			Resources repRes = ((XResForwarder) replacement).getResources();
+			int repId = ((XResForwarder) replacement).getId();
+			repRes.getValueForDensity(repId, density, outValue, resolveRefs);
+		} else {
+			if (replacement != null) {
+				XposedBridge.log("Replacement of resource ID #0x" + Integer.toHexString(id) + " escaped because of deprecated replacement. Please use XResForwarder instead.");
+			}
+			super.getValueForDensity(id, density, outValue, resolveRefs);
+		}
 	}
 
 	/** @hide */
@@ -1330,6 +1396,19 @@ public class XResources extends XResourcesSuperClass {
 		}
 
 		@Override
+		public Typeface getFont(int index) {
+			Object replacement = ((XResources) getResources()).getReplacement(getResourceId(index, 0));
+			if (replacement instanceof Typeface) {
+				return (Typeface) replacement;
+			} else if (replacement instanceof XResForwarder) {
+				Resources repRes = ((XResForwarder) replacement).getResources();
+				int repId = ((XResForwarder) replacement).getId();
+				return repRes.getFont(repId);
+			}
+			return super.getFont(index);
+		}
+
+		@Override
 		public float getFraction(int index, int base, int pbase, float defValue) {
 			Object replacement = ((XResources) getResources()).getReplacement(getResourceId(index, 0));
 			if (replacement instanceof XResForwarder) {
@@ -1426,6 +1505,44 @@ public class XResources extends XResourcesSuperClass {
 				return repRes.getTextArray(repId);
 			}
 			return super.getTextArray(index);
+		}
+
+		@Override
+		public boolean getValue(int index, TypedValue outValue) {
+			var id = getResourceId(index, 0);
+			Object replacement = ((XResources) getResources()).getReplacement(id);
+			if (replacement instanceof XResForwarder) {
+				Resources repRes = ((XResForwarder) replacement).getResources();
+				int repId = ((XResForwarder) replacement).getId();
+				repRes.getValue(repId, outValue, true);
+				return outValue.type != TypedValue.TYPE_NULL;
+			} else {
+				if (replacement != null) {
+					XposedBridge.log("Replacement of resource ID #0x" + Integer.toHexString(id) + " escaped because of deprecated replacement. Please use XResForwarder instead.");
+				}
+				return super.getValue(index, outValue);
+			}
+		}
+
+		@Override
+		public TypedValue peekValue(int index) {
+			var id = getResourceId(index, 0);
+			Object replacement = ((XResources) getResources()).getReplacement(id);
+			if (replacement instanceof XResForwarder) {
+				if (getBooleanField(this, "mRecycled")) {
+					throw new RuntimeException("Cannot make calls to a recycled instance!");
+				}
+				final TypedValue value = (TypedValue) getObjectField(this, "mValue");
+				Resources repRes = ((XResForwarder) replacement).getResources();
+				int repId = ((XResForwarder) replacement).getId();
+				repRes.getValue(repId, value, true);
+				return value;
+			} else {
+				if (replacement != null) {
+					XposedBridge.log("Replacement of resource ID #0x" + Integer.toHexString(id) + " escaped because of deprecated replacement. Please use XResForwarder instead.");
+				}
+				return super.peekValue(index);
+			}
 		}
 	}
 
@@ -1645,6 +1762,14 @@ public class XResources extends XResourcesSuperClass {
 	private static XC_LayoutInflated.Unhook hookLayoutInternal(String resDir, int id, ResourceNames resNames, XC_LayoutInflated callback) {
 		if (id == 0)
 			throw new IllegalArgumentException("id 0 is not an allowed resource identifier");
+
+		if (resDir == null) {
+			try {
+				XposedInit.hookResources();
+			} catch (Throwable throwable) {
+				throw new IllegalStateException("Failed to initialize resources hook", throwable);
+			}
+		}
 
 		HashMap<String, CopyOnWriteSortedSet<XC_LayoutInflated>> inner;
 		synchronized (sLayoutCallbacks) {
