@@ -1,23 +1,3 @@
-/*
- * This file is part of LSPosed.
- *
- * LSPosed is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * LSPosed is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with LSPosed.  If not, see <https://www.gnu.org/licenses/>.
- *
- * Copyright (C) 2020 EdXposed Contributors
- * Copyright (C) 2021 LSPosed Contributors
- */
-
 package de.robv.android.xposed;
 
 import android.app.Application;
@@ -43,26 +23,15 @@ public interface IXposedHookLoadPackage extends IXposedMod {
      */
     void handleLoadPackage(LoadPackageParam lpparam) throws Throwable;
 
-    /**
-     * @hide
-     */
+    /** @hide */
     final class Wrapper extends XC_LoadPackage {
         private final IXposedHookLoadPackage instance;
-        private final String apkPath;
-
-        public Wrapper(IXposedHookLoadPackage instance, String apkPath) {
+        public Wrapper(IXposedHookLoadPackage instance) {
             this.instance = instance;
-            this.apkPath = apkPath;
         }
-
         @Override
         public void handleLoadPackage(LoadPackageParam lpparam) throws Throwable {
             instance.handleLoadPackage(lpparam);
-        }
-
-        @Override
-        public String getApkPath() {
-            return apkPath;
         }
     }
 }
