@@ -22,17 +22,24 @@ package org.lsposed.lspd.hooker;
 
 import static org.lsposed.lspd.util.Utils.logD;
 
+import androidx.annotation.NonNull;
+
+import org.lsposed.lspd.impl.LSPosedContext;
 import org.lsposed.lspd.util.Hookers;
 
-import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XposedBridge;
 import de.robv.android.xposed.XposedInit;
 import de.robv.android.xposed.callbacks.XC_LoadPackage;
+import io.github.libxposed.api.XposedInterface;
+import io.github.libxposed.api.XposedModuleInterface;
+import io.github.libxposed.api.annotations.BeforeInvocation;
+import io.github.libxposed.api.annotations.XposedHooker;
 
-public class StartBootstrapServicesHooker extends XC_MethodHook {
+@XposedHooker
+public class StartBootstrapServicesHooker implements XposedInterface.Hooker {
 
-    @Override
-    protected void beforeHookedMethod(MethodHookParam param) {
+    @BeforeInvocation
+    public static void beforeHookedMethod() {
         logD("SystemServer#startBootstrapServices() starts");
 
         try {
@@ -41,10 +48,18 @@ public class StartBootstrapServicesHooker extends XC_MethodHook {
             XC_LoadPackage.LoadPackageParam lpparam = new XC_LoadPackage.LoadPackageParam(XposedBridge.sLoadedPackageCallbacks);
             lpparam.packageName = "android";
             lpparam.processName = "android"; // it's actually system_server, but other functions return this as well
-            lpparam.classLoader = SystemMainHooker.systemServerCL;
+            lpparam.classLoader = HandleSystemServerProcessHooker.systemServerCL;
             lpparam.appInfo = null;
             lpparam.isFirstApplication = true;
             XC_LoadPackage.callAll(lpparam);
+
+            LSPosedContext.callOnSystemServerLoaded(new XposedModuleInterface.SystemServerLoadedParam() {
+                @Override
+                @NonNull
+                public ClassLoader getClassLoader() {
+                    return HandleSystemServerProcessHooker.systemServerCL;
+                }
+            });
         } catch (Throwable t) {
             Hookers.logE("error when hooking startBootstrapServices", t);
         }
